+++ Dockerfile
@@ -10,23 +10,11 @@ RUN pip install -r requirements.txt
 ADD . .
 
 RUN yarn cache clean
-RUN cd front && \ 
+RUN cd frontend && \ 
     yarn install && \
     yarn build && \
     cd ..
 
-# EXPOSE 800
-# RUN cd mydarknet && \
-#     make clean && \
-#     make && \
-#     cd ..
+ENV PORT=8000
 
-# RUN python manage.py migrate
-
-# RUN python manage.py collectstatic --no-input
-
-
-
-
-# CMD gunicorn --bind 0.0.0.0:$PORT --workers 1 counting_vertebrae.wsgi:application
 ENTRYPOINT [ "./activate.sh" ]
