+++ counting_vertebrae/settings.py
@@ -21,11 +21,12 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 
 # SECURITY WARNING: keep the secret key used in production secret!
 import environ
+import os
 env = environ.Env()
 SECRET_KEY = "django-insecure-qii=6s!$pfuz(66)^6%j$x8dv+(=%mrcg+pa6=dmbo*3pbvd-_"
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
+DEBUG = env('DEBUG', False)
 
 ALLOWED_HOSTS = ['*']
 
@@ -75,7 +76,7 @@ TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
         'DIRS': [
-            BASE_DIR / 'front' / 'build'
+            BASE_DIR / 'frontend' / 'build'
         ],
         'APP_DIRS': True,
         'OPTIONS': {
@@ -89,7 +90,7 @@ TEMPLATES = [
     },
 ]
 STATICFILES_DIRS = [
-    BASE_DIR / 'front' / 'build' / 'static'
+    BASE_DIR / 'frontend' / 'build' / 'static'
 ]
 WSGI_APPLICATION = 'counting_vertebrae.wsgi.application'
 
@@ -149,9 +150,9 @@ STATIC_ROOT = BASE_DIR / 'staticfiles'
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 
-CELERY_BROKER_URL = "redis://:p3bfe17c51ffcaed5a595a89ee0e5230682c604afed54460ea7eceac99eabf6e3@ec2-107-20-253-248.compute-1.amazonaws.com:17700"
+# CELERY_BROKER_URL = "redis://:p3bfe17c51ffcaed5a595a89ee0e5230682c604afed54460ea7eceac99eabf6e3@ec2-107-20-253-248.compute-1.amazonaws.com:17700"
 
-CELERY_RESULT_BACKEND = "redis://:p3bfe17c51ffcaed5a595a89ee0e5230682c604afed54460ea7eceac99eabf6e3@ec2-107-20-253-248.compute-1.amazonaws.com:17700"
+# CELERY_RESULT_BACKEND = "redis://:p3bfe17c51ffcaed5a595a89ee0e5230682c604afed54460ea7eceac99eabf6e3@ec2-107-20-253-248.compute-1.amazonaws.com:17700"
 
 # CELERY_ACCEPT_CONTENT = ['application/json']
 
@@ -160,6 +161,3 @@ CELERY_RESULT_BACKEND = "redis://:p3bfe17c51ffcaed5a595a89ee0e5230682c604afed544
 # CELERY_TASK_SERIALIZER = 'json'
 
 
-
-
-# compute-instance-001.uintl8.0001.usw2.cache.amazonaws.com
