+++ counting_vertebrae/urls.py
@@ -8,7 +8,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('detect/', include('detector.urls'), name="detector"),
     path('', TemplateView.as_view(template_name='index.html'), name='index'),
-    # re_path(r"\S", TemplateView.as_view(template_name="index.html")),
 ]
 
 # start_listening()
