--- detector/tasks.py
@@ -1,5 +1,0 @@
-from celery import shared_task
-
-@shared_task
-def add(x, y):
-    return x + y
