+++ detector/urls.py
@@ -4,5 +4,4 @@ from .views import ImageDetectorView, ImageDetectorReturnImage
 urlpatterns = [
     path('', ImageDetectorView.as_view()),
     path('image/', ImageDetectorReturnImage.as_view())
-
 ]
