+++ docker-compose.yml
@@ -14,26 +14,5 @@ services:
       - .env
         
     command: ["./activate.sh"]
-
-  # frontend:
-
-
-
-  
-  # subscriber:
-  #   build: 
-  #     context: .
-  #     dockerfile: Dockerfile
-
-  #   env_file: 
-  #     - .env
-    
-  #   volumes:
-  #     - './mydarknet:/api/mydarknet'
-    
-
-  #   restart: "on-failure"
-
-  #   command: python subscriber.py
