+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "front",
+  "name": "vertebrae-counter",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
