+++ frontend/src/components/Detections/Detections.js
