+++ frontend/src/components/ImageDisplay/options.js
