--- instructions/YOLO_ Real-Time Object Detection_files/ad_status.js
@@ -1,1 +1,0 @@
-window.google_ad_status = 1;
