--- instructions/YOLO_ Real-Time Object Detection_files/embed.js
@@ -1,86 +1,0 @@
-(function(g){var window=this;var iua=function(a,b){var c=(b-a.o)/(a.u-a.o);if(0>=c)return 0;if(1<=c)return 1;for(var d=0,e=1,f=0,k=0;8>k;k++){f=g.Tn(a,c);var l=(g.Tn(a,c+1E-6)-f)/1E-6;if(1E-6>Math.abs(f-b))return c;if(1E-6>Math.abs(l))break;else f<b?d=c:e=c,c-=(f-b)/l}for(k=0;1E-6<Math.abs(f-b)&&8>k;k++)f<b?(d=c,c=(c+e)/2):(e=c,c=(c+d)/2),f=g.Tn(a,c);return c},jua=function(){return g.X?{D:"div",
-Y:["ytp-icon","ytp-icon-small-close"]}:{D:"svg",O:{height:"100%",viewBox:"0 0 16 16",width:"100%"},K:[{D:"path",O:{d:"M13 4L12 3 8 7 4 3 3 4 7 8 3 12 4 13 8 9 12 13 13 12 9 8z",fill:"#fff"}}]}},F4=function(){return g.X?{D:"div",
-Y:["ytp-icon","ytp-icon-watermark"]}:{D:"svg",O:{height:"100%",version:"1.1",viewBox:"0 0 110 26",width:"100%"},K:[{D:"path",La:!0,I:"ytp-svg-fill",O:{d:"M 16.68,.99 C 13.55,1.03 7.02,1.16 4.99,1.68 c -1.49,.4 -2.59,1.6 -2.99,3 -0.69,2.7 -0.68,8.31 -0.68,8.31 0,0 -0.01,5.61 .68,8.31 .39,1.5 1.59,2.6 2.99,3 2.69,.7 13.40,.68 13.40,.68 0,0 10.70,.01 13.40,-0.68 1.5,-0.4 2.59,-1.6 2.99,-3 .69,-2.7 .68,-8.31 .68,-8.31 0,0 .11,-5.61 -0.68,-8.31 -0.4,-1.5 -1.59,-2.6 -2.99,-3 C 29.11,.98 18.40,.99 18.40,.99 c 0,0 -0.67,-0.01 -1.71,0 z m 72.21,.90 0,21.28 2.78,0 .31,-1.37 .09,0 c .3,.5 .71,.88 1.21,1.18 .5,.3 1.08,.40 1.68,.40 1.1,0 1.99,-0.49 2.49,-1.59 .5,-1.1 .81,-2.70 .81,-4.90 l 0,-2.40 c 0,-1.6 -0.11,-2.90 -0.31,-3.90 -0.2,-0.89 -0.5,-1.59 -1,-2.09 -0.5,-0.4 -1.10,-0.59 -1.90,-0.59 -0.59,0 -1.18,.19 -1.68,.49 -0.49,.3 -1.01,.80 -1.21,1.40 l 0,-7.90 -3.28,0 z m -49.99,.78 3.90,13.90 .18,6.71 3.31,0 0,-6.71 3.87,-13.90 -3.37,0 -1.40,6.31 c -0.4,1.89 -0.71,3.19 -0.81,3.99 l -0.09,0 c -0.2,-1.1 -0.51,-2.4 -0.81,-3.99 l -1.37,-6.31 -3.40,0 z m 29.59,0 0,2.71 3.40,0 0,17.90 3.28,0 0,-17.90 3.40,0 c 0,0 .00,-2.71 -0.09,-2.71 l -9.99,0 z m -53.49,5.12 8.90,5.18 -8.90,5.09 0,-10.28 z m 89.40,.09 c -1.7,0 -2.89,.59 -3.59,1.59 -0.69,.99 -0.99,2.60 -0.99,4.90 l 0,2.59 c 0,2.2 .30,3.90 .99,4.90 .7,1.1 1.8,1.59 3.5,1.59 1.4,0 2.38,-0.3 3.18,-1 .7,-0.7 1.09,-1.69 1.09,-3.09 l 0,-0.5 -2.90,-0.21 c 0,1 -0.08,1.6 -0.28,2 -0.1,.4 -0.5,.62 -1,.62 -0.3,0 -0.61,-0.11 -0.81,-0.31 -0.2,-0.3 -0.30,-0.59 -0.40,-1.09 -0.1,-0.5 -0.09,-1.21 -0.09,-2.21 l 0,-0.78 5.71,-0.09 0,-2.62 c 0,-1.6 -0.10,-2.78 -0.40,-3.68 -0.2,-0.89 -0.71,-1.59 -1.31,-1.99 -0.7,-0.4 -1.48,-0.59 -2.68,-0.59 z m -50.49,.09 c -1.09,0 -2.01,.18 -2.71,.68 -0.7,.4 -1.2,1.12 -1.49,2.12 -0.3,1 -0.5,2.27 -0.5,3.87 l 0,2.21 c 0,1.5 .10,2.78 .40,3.78 .2,.9 .70,1.62 1.40,2.12 .69,.5 1.71,.68 2.81,.78 1.19,0 2.08,-0.28 2.78,-0.68 .69,-0.4 1.09,-1.09 1.49,-2.09 .39,-1 .49,-2.30 .49,-3.90 l 0,-2.21 c 0,-1.6 -0.2,-2.87 -0.49,-3.87 -0.3,-0.89 -0.8,-1.62 -1.49,-2.12 -0.7,-0.5 -1.58,-0.68 -2.68,-0.68 z m 12.18,.09 0,11.90 c -0.1,.3 -0.29,.48 -0.59,.68 -0.2,.2 -0.51,.31 -0.81,.31 -0.3,0 -0.58,-0.10 -0.68,-0.40 -0.1,-0.3 -0.18,-0.70 -0.18,-1.40 l 0,-10.99 -3.40,0 0,11.21 c 0,1.4 .18,2.39 .68,3.09 .49,.7 1.21,1 2.21,1 1.4,0 2.48,-0.69 3.18,-2.09 l .09,0 .31,1.78 2.59,0 0,-14.99 c 0,0 -3.40,.00 -3.40,-0.09 z m 17.31,0 0,11.90 c -0.1,.3 -0.29,.48 -0.59,.68 -0.2,.2 -0.51,.31 -0.81,.31 -0.3,0 -0.58,-0.10 -0.68,-0.40 -0.1,-0.3 -0.21,-0.70 -0.21,-1.40 l 0,-10.99 -3.40,0 0,11.21 c 0,1.4 .21,2.39 .71,3.09 .5,.7 1.18,1 2.18,1 1.39,0 2.51,-0.69 3.21,-2.09 l .09,0 .28,1.78 2.62,0 0,-14.99 c 0,0 -3.40,.00 -3.40,-0.09 z m 20.90,2.09 c .4,0 .58,.11 .78,.31 .2,.3 .30,.59 .40,1.09 .1,.5 .09,1.21 .09,2.21 l 0,1.09 -2.5,0 0,-1.09 c 0,-1 -0.00,-1.71 .09,-2.21 0,-0.4 .11,-0.8 .31,-1 .2,-0.3 .51,-0.40 .81,-0.40 z m -50.49,.12 c .5,0 .8,.18 1,.68 .19,.5 .28,1.30 .28,2.40 l 0,4.68 c 0,1.1 -0.08,1.90 -0.28,2.40 -0.2,.5 -0.5,.68 -1,.68 -0.5,0 -0.79,-0.18 -0.99,-0.68 -0.2,-0.5 -0.31,-1.30 -0.31,-2.40 l 0,-4.68 c 0,-1.1 .11,-1.90 .31,-2.40 .2,-0.5 .49,-0.68 .99,-0.68 z m 39.68,.09 c .3,0 .61,.10 .81,.40 .2,.3 .27,.67 .37,1.37 .1,.6 .12,1.51 .12,2.71 l .09,1.90 c 0,1.1 .00,1.99 -0.09,2.59 -0.1,.6 -0.19,1.08 -0.49,1.28 -0.2,.3 -0.50,.40 -0.90,.40 -0.3,0 -0.51,-0.08 -0.81,-0.18 -0.2,-0.1 -0.39,-0.29 -0.59,-0.59 l 0,-8.5 c .1,-0.4 .29,-0.7 .59,-1 .3,-0.3 .60,-0.40 .90,-0.40 z"}}]}},
-kua=function(){return g.X?{D:"div",
-Y:["ytp-icon","ytp-icon-youtube-logo-redirect"]}:{D:"svg",O:{fill:"#fff",height:"100%",viewBox:"0 0 24 24",width:"100%"},K:[{D:"path",O:{d:"M0 0h24v24H0V0z",fill:"none"}},{D:"path",O:{d:"M21.58 7.19c-.23-.86-.91-1.54-1.77-1.77C18.25 5 12 5 12 5s-6.25 0-7.81.42c-.86.23-1.54.91-1.77 1.77C2 8.75 2 12 2 12s0 3.25.42 4.81c.23.86.91 1.54 1.77 1.77C5.75 19 12 19 12 19s6.25 0 7.81-.42c.86-.23 1.54-.91 1.77-1.77C22 15.25 22 12 22 12s0-3.25-.42-4.81zM10 15V9l5.2 3-5.2 3z"}}]}},G4=function(a){g.A.call(this);
-this.B=a;this.A=new g.Sn(0,0,.4,0,.2,1,1,1);this.u=new g.wn(this.w,window,this);g.C(this,this.u)},H4=function(a){g.Q.call(this,{D:"div",
-I:"ytp-related-on-error-overlay"});this.R=this.A=0;this.u=a;this.Z=new g.O(this);g.C(this,this.Z);a=a.N();this.C=[];this.w=[];this.F=0;this.ca=a.u;this.M=new g.Q({D:"h2",I:"ytp-related-title",W:"{{title}}"});g.C(this,this.M);this.M.ba(this.element);this.G=new g.Q({D:"div",I:"ytp-suggestions"});g.C(this,this.G);this.G.ba(this.element);this.J=new g.Q({D:"button",Y:["ytp-button","ytp-previous"],O:{"aria-label":"Show previous suggested videos"},K:[g.SL()]});g.C(this,this.J);this.J.ba(this.element);this.J.ia("click",
-this.ZH,this);this.da=new G4((0,g.x)(this.gz,this));g.C(this,this.da);this.aa=this.B=0;this.T=!0;for(var b=0;16>b;b++){var c=new g.Q({D:"a",I:"ytp-suggestion-link",O:{href:"{{link}}",target:a.A,"aria-label":"{{aria_label}}"},K:[{D:"div",I:"ytp-suggestion-image",K:[{D:"div",O:{"data-is-live":"{{is_live}}"},I:"ytp-suggestion-duration",W:"{{duration}}"}]},{D:"div",I:"ytp-suggestion-title",O:{title:"{{hover_title}}"},W:"{{title}}"},{D:"div",I:"ytp-suggestion-author",W:"{{views_or_author}}"}]});g.C(this,
-c);c.ba(this.G.element);var d=c.o["ytp-suggestion-link"];g.vh(d,"transitionDelay",b/20+"s");this.Z.L(d,"click",g.Sa(this.aI,b));this.C.push(c)}this.H=new g.Q({D:"button",Y:["ytp-button","ytp-next"],O:{"aria-label":"Show more suggested videos"},K:[g.TL()]});g.C(this,this.H);this.H.ba(this.element);this.H.ia("click",this.YH,this);this.Z.L(this.u,"videodatachange",this.fz);lua(this,g.NK(this.u).getPlayerSize());this.fz();this.show()},lua=function(a,b,c){var d=a.u.N(),e=16/9,f=650<=b.width,k=480>b.width||
-290>b.height,l=Math.min(a.w.length,a.C.length);
-if(150>=Math.min(b.width,b.height)||0==l||!d.Ha)a.hide();else{var m;if(f){var n=m=28;a.A=16}else n=m=8,a.A=8;if(k){var p=6;f=14;var t=12;k=24;d=12}else p=8,f=18,t=16,k=36,d=16;b=b.width-(48+m+n);m=Math.ceil(b/150);m=Math.min(3,m);m=b/m-a.A;n=Math.floor(m/e);c&&n+100>c&&50<m&&(n=Math.max(c,50/e),m=Math.ceil(b/(e*(n-100)+a.A)),m=b/m-a.A,n=Math.floor(m/e));50>m||g.PK(a.u)?a.hide():a.show();for(c=0;c<l;c++){e=a.C[c];var u=e.o["ytp-suggestion-image"];u.style.width=m+"px";u.style.height=n+"px";e.o["ytp-suggestion-title"].style.width=
-m+"px";e.o["ytp-suggestion-author"].style.width=m+"px";e=e.o["ytp-suggestion-duration"];e.style.display=e&&100>m?"none":""}l=f+p+t+4;a.R=l+d+(n-k)/2;a.G.element.style.height=n+l+"px";a.aa=m;a.F=b;a.B=0;a.gz(0);I4(a)}},mua=function(a,b){var c=g.Od(b,a.F-a.w.length*(a.aa+a.A),0);
-a.da.start(a.B,c,1E3);a.B=c;I4(a)},I4=function(a){a.H.element.style.bottom=a.R+"px";
-a.J.element.style.bottom=a.R+"px";var b=a.B,c=a.F-a.w.length*(a.aa+a.A);g.K(a.element,"ytp-scroll-min",0<=b);g.K(a.element,"ytp-scroll-max",b<=c)},nua=function(a){for(var b=0;b<a.w.length;b++){var c=b,d=a.w[b].va();
-c=a.C[c];var e=d.shortViewCount?d.shortViewCount:d.author,f=d.mi();g.Iy(a.u.N())&&(f=g.Ad(f,g.aD({},"emb_rel_err")));c.element.style.display="";var k=c.o["ytp-suggestion-title"];g.fn.test(d.title)?k.dir="rtl":g.xqa.test(d.title)&&(k.dir="ltr");k=c.o["ytp-suggestion-author"];g.fn.test(e)?k.dir="rtl":g.xqa.test(e)&&(k.dir="ltr");k=void 0;d.ua?k="Live":k=d.lengthSeconds?g.KL(d.lengthSeconds):"";c.update({views_or_author:e,duration:k,link:f,hover_title:d.title,title:d.title,aria_label:d.Ak||null,is_live:d.ua});
-c=c.o["ytp-suggestion-image"];d=d.zc();c.style.backgroundImage=d?"url("+d+")":""}for(;b<a.C.length;b++)a.C[b].element.style.display="none";I4(a)},J4=function(a){g.EM.call(this,a);
-a=a.N();a=new g.Q({D:"a",I:"ytp-small-redirect",O:{href:g.pz(a),target:a.A,"aria-label":"Visit YouTube to search for more videos"},K:[kua()]});a.ba(this.element);g.C(this,a);this.w=new H4(this.api);this.w.ba(this.element);g.C(this,this.w);this.C=1;this.u(g.NK(this.api).getPlayerSize())},oua=function(a,b){a.o["ytp-error-content"].style.paddingTop="0px";
-var c=a.o["ytp-error-content"],d=c.clientHeight;lua(a.w,b,b.height-d);if(a.api.N().B){var e=g.$L(),f=g.X?{D:"div",Y:["ytp-icon","ytp-icon-youtube-logo-redirect-large"]}:{D:"svg",O:{fill:"#fff",viewBox:"0 0 24 24"},K:[{D:"path",O:{d:"M0 0h24v24H0V0z",fill:"none"}},{D:"path",O:{d:"M21.58 7.19c-.23-.86-.91-1.54-1.77-1.77C18.25 5 12 5 12 5s-6.25 0-7.81.42c-.86.23-1.54.91-1.77 1.77C2 8.75 2 12 2 12s0 3.25.42 4.81c.23.86.91 1.54 1.77 1.77C5.75 19 12 19 12 19s6.25 0 7.81-.42c.86-.23 1.54-.91 1.77-1.77C22 15.25 22 12 22 12s0-3.25-.42-4.81zM10 15V9l5.2 3-5.2 3z"}}]},
-k=1;480>b.width||290>b.height?(e={D:"div",Y:["ytp-icon","ytp-icon-error-exclamation-small"]},f=kua(),k=0):650<=b.width&&(e={D:"div",Y:["ytp-icon","ytp-icon-error-exclamation-large"]},f=g.X?{D:"div",Y:["ytp-icon","ytp-icon-youtube-logo-redirect-extra-large"]}:{D:"svg",O:{fill:"#fff",height:"100%",viewBox:"0 0 24 24",width:"100%"},K:[{D:"path",O:{d:"M0 0h24v24H0V0z",fill:"none"}},{D:"path",O:{d:"M21.58 7.19c-.23-.86-.91-1.54-1.77-1.77C18.25 5 12 5 12 5s-6.25 0-7.81.42c-.86.23-1.54.91-1.77 1.77C2 8.75 2 12 2 12s0 3.25.42 4.81c.23.86.91 1.54 1.77 1.77C5.75 19 12 19 12 19s6.25 0 7.81-.42c.86-.23 1.54-.91 1.77-1.77C22 15.25 22 12 22 12s0-3.25-.42-4.81zM10 15V9l5.2 3-5.2 3z"}}]},
-k=2);if(k!==a.C){e={"ytp-error-icon-container":e,"ytp-small-redirect":f};f=g.q(Object.keys(e));for(var l=f.next();!l.done;l=f.next()){l=l.value;var m=g.ee(l,a.element);g.te(m);(new g.Q(e[l])).ba(m)}a.C=k}}c.style.paddingTop=(b.height-a.w.element.clientHeight)/2-d/2+"px"},M4=function(a,b){var c=a.N();
-g.Q.call(this,{D:"button",Y:["ytp-impression-link","ytp-button"]});this.hide();this.u=a;if(!c.sa&&!b.Qd()){g.R(c.experiments,"embeds_impression_link_call_to_action")&&(g.I(this.element,"show-cta-button"),(new g.Q({D:"div",I:"ytp-impression-link-content",K:[{D:"div",I:"ytp-impression-link-text",W:"Watch on"},{D:"div",I:"ytp-impression-link-logo",K:[F4()]}]})).ba(this.element),this.show());if(g.R(c.experiments,"embeds_impression_link_video_thumbnail")){g.I(this.element,"show-video-thumbnail-button");
-(new g.Q({D:"div",I:"ytp-impression-link-header",W:"More from YouTube"})).ba(this.element);var d=this.u.getVideoData();d=new g.Q({D:"div",I:"ytp-impression-link-content",K:[{D:"div",I:"ytp-impression-link-metadata",K:[{D:"div",I:"ytp-impression-link-title",W:d.title},{D:"div",I:"ytp-impression-link-views-and-duration",W:"{{views_and_duration}}"}]},{D:"div",I:"ytp-impression-link-thumbnail"}]});d.ba(this.element);K4(this,d);L4(this,d);this.show()}g.R(c.experiments,"embeds_impression_link_channel_thumbnail")&&
-(g.I(this.element,"show-channel-thumbnail-button"),(new g.Q({D:"div",I:"ytp-impression-link-header",W:"More from YouTube"})).ba(this.element),d=this.u.getVideoData(),d=new g.Q({D:"div",I:"ytp-impression-link-content",K:[{D:"div",I:"ytp-impression-link-metadata",K:[{D:"div",I:"ytp-impression-link-title",W:d.Qk},{D:"div",I:"ytp-impression-link-subscribers",W:d.expandedSubtitle}]},{D:"div",I:"ytp-impression-link-thumbnail"}]}),d.ba(this.element),K4(this,d),this.show());g.R(c.experiments,"embeds_impression_link_occlusion")&&
-pua(this);g.R(c.experiments,"embeds_impression_link_hover")&&(qua(this),this.show());this.L(a,"presentingplayerstatechange",this.A);this.L(this.element,"click",this.w)}},pua=function(a){function b(d){g.q(d).next().value.isIntersecting||a.show()}
-g.I(a.element,"show-occlusion-video-thumbnail-button");(new g.Q({D:"div",I:"ytp-impression-link-header",W:"More from YouTube"})).ba(a.element);var c=a.u.getVideoData();c=new g.Q({D:"div",I:"ytp-impression-link-content",K:[{D:"div",I:"ytp-impression-link-metadata",K:[{D:"div",I:"ytp-impression-link-title",W:c.title},{D:"div",I:"ytp-impression-link-author",W:c.Qk},{D:"div",I:"ytp-impression-link-views",W:"{{views}}"}]},{D:"div",I:"ytp-impression-link-thumbnail-and-duration",K:[{D:"div",I:"ytp-impression-link-thumbnail"},
-{D:"div",I:"ytp-impression-link-duration",W:"{{duration}}"}]}]});c.ba(a.element);K4(a,c);L4(a,c);c=new g.Q({D:"button",Y:["ytp-button","ytp-impression-link-close"],K:[{D:"div",Y:["ytp-impression-link-close-icon"],K:[jua()]}]});c.ba(a.element);c.ia("click",function(d){a.hide();d.stopPropagation()},a);
-try{(new IntersectionObserver(b,{threshold:.8})).observe(document.querySelector("body"))}catch(d){g.M(d)}},qua=function(a){g.I(a.element,"show-video-thumbnail-expanding-button");
-a.L(a.element,"mouseenter",function(){g.I(a.element,"show-expanded-metadata");g.Ln(a.element,"show-collapsed-metadata")});
-a.L(a.element,"mouseleave",function(){g.Ln(a.element,"show-expanded-metadata");g.I(a.element,"show-collapsed-metadata")});
-(new g.Q({D:"div",I:"ytp-impression-link-header",W:"More videos"})).ba(a.element);var b=a.u.getVideoData();b=new g.Q({D:"div",I:"ytp-impression-link-content",K:[{D:"div",I:"ytp-impression-link-metadata",K:[{D:"div",I:"ytp-impression-link-title",W:b.title},{D:"div",I:"ytp-impression-link-views-and-duration",W:"{{views_and_duration}}"}]},{D:"div",I:"ytp-impression-link-thumbnail"}]});b.ba(a.element);K4(a,b);L4(a,b)},K4=function(a,b){var c=a.u.N(),d=a.u.getVideoData();
-c=g.R(c.experiments,"embeds_impression_link_channel_thumbnail")?d.Jd:d.zc();b.o["ytp-impression-link-thumbnail"].style.backgroundImage=c?"url("+c+")":""},L4=function(a,b){var c=a.u.N(),d=a.u.getVideoData(),e="";
-d.ua?e="Live":d.lengthSeconds&&(e=g.KL(d.lengthSeconds));d=d.shortViewCount?d.shortViewCount:"";var f="";d&&e?f=d+" \u2022 "+e:d?f=d:e&&(f=e);g.R(c.experiments,"embeds_impression_link_occlusion")?b.update({views:d,duration:e}):b.update({views_and_duration:f})},N4=function(a){g.Q.call(this,{D:"div",
-I:"ytp-muted-autoplay-endscreen-overlay",K:[{D:"div",I:"ytp-muted-autoplay-end-panel",K:[{D:"div",I:"ytp-muted-autoplay-end-text",W:"{{text}}"}]}]});this.u=a;this.A=this.o["ytp-muted-autoplay-end-panel"];this.w=new g.rM(a);g.C(this,this.w);this.w.ba(this.A,0);g.hL(a,this.element,this,52428);this.L(a,"presentingplayerstatechange",this.C);this.ia("click",this.B);this.hide()},O4=function(a){g.Q.call(this,{D:"div",
-I:"ytp-muted-autoplay-overlay",K:[{D:"div",I:"ytp-muted-autoplay-bottom-buttons",K:[{D:"button",Y:["ytp-muted-autoplay-equalizer","ytp-button"],K:[{D:"div",Y:["ytp-muted-autoplay-equalizer-icon"],K:[g.X?{D:"div",Y:["ytp-icon","ytp-icon-equalizer-animation"]}:{D:"svg",O:{height:"100%",version:"1.1",viewBox:"-4 -4 24 24",width:"100%"},K:[{D:"g",O:{fill:"#fff"},K:[{D:"rect",I:"ytp-equalizer-bar-left",O:{height:"9",width:"4",x:"1",y:"7"}},{D:"rect",I:"ytp-equalizer-bar-middle",O:{height:"14",width:"4",
-x:"6",y:"2"}},{D:"rect",I:"ytp-equalizer-bar-right",O:{height:"12",width:"4",x:"11",y:"4"}}]}]}]}]}]}]});this.u=a;this.bottomButtons=this.o["ytp-muted-autoplay-bottom-buttons"];this.A=this.o["ytp-muted-autoplay-equalizer"];this.B=new g.H(this.C,4E3,this);g.hL(a,this.element,this,39306);this.L(a,"presentingplayerstatechange",this.w);this.L(a,"onMutedAutoplayStarts",this.w);this.ia("click",this.F);this.hide()},P4=function(a,b){g.Q.call(this,{D:"div",
-I:"ytp-pause-overlay"});this.u=a;this.ca=b;this.A=new g.O(this);g.C(this,this.A);this.T=new g.OL(this,1E3,!1,100);g.C(this,this.T);var c=a.N();"0"==c.controlsType&&g.I(a.getRootNode(),"ytp-pause-overlay-controls-hidden");this.C=[];this.w=[];this.F=0;this.da=c.u;this.aa=!1;this.na=0;this.R=new g.Q({D:"h2",I:"ytp-related-title",W:"{{title}}"});g.C(this,this.R);this.R.ba(this.element);this.G=new g.Q({D:"div",I:"ytp-suggestions"});g.C(this,this.G);this.G.ba(this.element);this.M=new g.Q({D:"button",Y:["ytp-button",
-"ytp-previous"],O:{"aria-label":"Show previous suggested videos"},K:[g.SL()]});g.C(this,this.M);this.M.ba(this.element);this.M.ia("click",this.UH,this);var d=g.ax||g.yh?{style:"will-change: opacity"}:null;this.ea=new G4((0,g.x)(this.ez,this));g.C(this,this.ea);for(var e=this.ha=this.Z=this.B=0;16>e;e++){var f=new g.Q({D:"a",I:"ytp-suggestion-link",O:{href:"{{link}}",target:c.A,"aria-label":"{{aria_label}}"},K:[{D:"div",I:"ytp-suggestion-image"},{D:"div",I:"ytp-suggestion-overlay",O:d,K:[{D:"div",
-I:"ytp-suggestion-title",W:"{{title}}"},{D:"div",I:"ytp-suggestion-author",W:"{{author_and_views}}"},{D:"div",O:{"data-is-live":"{{is_live}}"},I:"ytp-suggestion-duration",W:"{{duration}}"}]}]});g.C(this,f);f.ba(this.G.element);var k=f.o["ytp-suggestion-link"];g.vh(k,"transitionDelay",e/20+"s");this.A.L(k,"click",g.Sa(this.VH,e));this.C.push(f)}this.J=new g.Q({D:"button",Y:["ytp-button","ytp-next"],O:{"aria-label":"Show more suggested videos"},K:[g.TL()]});g.C(this,this.J);this.J.ba(this.element);
-this.J.ia("click",this.TH,this);c=new g.Q({D:"button",Y:["ytp-button","ytp-collapse"],O:{"aria-label":"Hide more videos"},K:[jua()]});g.C(this,c);c.ba(this.element);c.ia("click",this.HK,this);this.H=new g.Q({D:"button",Y:["ytp-button","ytp-expand"],W:"More videos"});g.C(this,this.H);this.H.ba(this.element);this.H.ia("click",this.IK,this);this.A.L(this.u,"appresize",this.it);this.A.L(this.u,"fullscreentoggled",this.WH);this.A.L(this.u,"presentingplayerstatechange",this.XH);this.A.L(this.u,"videodatachange",
-this.dz);this.it(g.NK(this.u).getPlayerSize());this.dz()},rua=function(a,b){var c=g.Od(b,a.F-a.w.length*(a.Z+8),0);
-a.ea.start(a.B,c,1E3);a.B=c;Q4(a)},Q4=function(a){var b=a.ca.qc();
-b=a.ha/2+(b?32:16);a.J.element.style.bottom=b+"px";a.M.element.style.bottom=b+"px";b=a.B;var c=a.F-a.w.length*(a.Z+8);g.K(a.element,"ytp-scroll-min",0<=b);g.K(a.element,"ytp-scroll-max",b<=c)},sua=function(a){for(var b=0;b<a.w.length;b++){var c=a.w[b].va(),d=a.C[b],e=c.shortViewCount?c.author+" \u2022 "+c.shortViewCount:c.author;
-d.element.style.display="";g.wqa.test(c.title)&&(d.o["ytp-suggestion-title"].dir="rtl");g.wqa.test(e)&&(d.o["ytp-suggestion-author"].dir="rtl");var f=void 0;c.ua?f="Live":f=c.lengthSeconds?g.KL(c.lengthSeconds):"";var k=c.mi();g.Iy(a.u.N())&&(k=g.Ad(k,g.aD({},"emb_rel_pause")));d.update({author_and_views:e,duration:f,link:k,title:c.title,aria_label:c.Ak||null,is_live:c.ua});d=d.o["ytp-suggestion-image"];c=c.zc();d.style.backgroundImage=c?"url("+c+")":""}for(;b<a.C.length;b++)a.C[b].element.style.display=
-"none";Q4(a)},R4=function(a){var b=a.N();
-g.Q.call(this,{D:"a",Y:["ytp-watermark","yt-uix-sessionlink"],O:{target:b.A,href:"{{url}}","aria-label":g.tM("Watch on $WEBSITE",{WEBSITE:g.bz(b)}),"data-sessionlink":"feature=player-watermark"},K:[F4()]});this.u=a;this.A=null;this.B=!1;this.w=g.MK(a);g.hL(a,this.element,this,76758);this.L(a,"videodatachange",this.jt);this.L(a,"videodatachange",this.iz);this.L(a,"presentingplayerstatechange",this.cI);this.L(a,"appresize",this.hz);b=this.w;this.w!=b&&(this.w=b);this.jt();this.iz();this.hz(g.NK(a).getPlayerSize())},
-S4=function(a){g.sL.call(this,a);
-this.o=a;this.B=new g.O(this);g.C(this,this.B);this.load()},tua=function(a,b){g.V(b,128)?(a.u||(a.u=new J4(a.o),g.C(a,a.u),g.fL(a.o,a.u.element,4)),a.u.B(b.u),a.u.show(),g.I(a.o.getRootNode(),"ytp-embed-error")):a.u&&(a.u.dispose(),a.u=null,g.Ln(a.o.getRootNode(),"ytp-embed-error"))};
-g.r(G4,g.A);G4.prototype.start=function(a,b,c){this.o=a;this.G=b;this.C=c;this.F=(0,g.N)();this.w()};
-G4.prototype.w=function(){var a=(0,g.N)()-this.F;var b=this.A;a=iua(b,a/this.C);if(0==a)b=b.B;else if(1==a)b=b.G;else{var c=g.Qd(b.B,b.C,a),d=g.Qd(b.C,b.F,a);b=g.Qd(b.F,b.G,a);c=g.Qd(c,d,a);d=g.Qd(d,b,a);b=g.Qd(c,d,a)}b=g.Od(b,0,1);this.B((this.G-this.o)*b+this.o);1>b&&this.u.start()};g.r(H4,g.Q);g.h=H4.prototype;g.h.hide=function(){this.T=!0;g.Q.prototype.hide.call(this)};
-g.h.show=function(){this.T=!1;g.Q.prototype.show.call(this)};
-g.h.isHidden=function(){return this.T};
-g.h.YH=function(){mua(this,this.B-this.F)};
-g.h.ZH=function(){mua(this,this.B+this.F)};
-g.h.aI=function(a,b){var c=this.w[a],d=c.Lb;if(g.zM(b,this.u,this.ca,d||void 0)){var e=c.va().videoId;c=c.getPlaylistId();g.oT(this.u.app,e,d,c,void 0,void 0)}};
-g.h.fz=function(){var a=this,b=this.u.getVideoData(),c=this.u.N();this.ca=b.Ac?!1:c.u;if(b.suggestions){var d=(0,g.we)(b.suggestions,function(e){return e&&!e.list});
-this.w=(0,g.Ec)(d,function(e){e=g.VN(c,e);g.C(a,e);return e})}else this.w.length=0;
-nua(this);b.Ac?this.M.update({title:g.tM("More videos from $DNI_RELATED_CHANNEL",{DNI_RELATED_CHANNEL:b.author})}):this.M.update({title:"More videos on YouTube"})};
-g.h.gz=function(a){this.G.element.scrollLeft=-a};g.r(J4,g.EM);J4.prototype.show=function(){g.EM.prototype.show.call(this);oua(this,g.NK(this.api).getPlayerSize())};
-J4.prototype.u=function(a){g.EM.prototype.u.call(this,a);oua(this,a);g.K(this.element,"related-on-error-overlay-visible",!this.w.isHidden())};
-J4.prototype.B=function(a){g.EM.prototype.B.call(this,a);var b=this.api.getVideoData(),c=g.R(this.api.N().experiments,"embeds_enable_sg_misinfo"),d;if(a.Xg){if(b.Pn){a:{a=b.Pn;if(a.runs)for(var e=0;e<a.runs.length;e++)if(a.runs[e].navigationEndpoint){a=!0;break a}a=!1}a?d=g.Yz(b.Pn):d=g.FM(g.U(b.Pn))}else d=g.FM(a.Xg);this.kb(d,"subreason")}(b=b.Kw)&&c&&(b.reason&&(c=g.Yz(b.reason),this.kb(c,"content")),b.subreason&&(c=g.Yz(b.subreason),this.kb(c,"subreason")))};g.r(M4,g.Q);M4.prototype.A=function(){g.MK(this.u).isCued()||this.hide()};
-M4.prototype.w=function(a){var b=this.u.N(),c=this.u.getVideoData();b=g.R(b.experiments,"embeds_impression_link_channel_thumbnail")?g.pz(b)+c.kf:this.u.getVideoUrl();g.AM(b,this.u,a)};g.r(N4,g.Q);N4.prototype.C=function(){var a=g.MK(this.u),b=this.u.getVideoData();g.R(this.u.N().experiments,"embeds_enable_muted_autoplay")&&b.mutedAutoplay&&(g.V(a,2)&&!this.Fa()?(this.show(),this.w.show(),a=this.u.getVideoData(),this.ka("text",a.Zz),g.K(this.element,"ytp-muted-autoplay-show-end-panel",!0),g.lL(this.u,this.element,this.Fa()),this.u.ma("onMutedAutoplayEnds")):this.hide())};
-N4.prototype.B=function(){var a=this.u.getVideoData(),b=this.u.getCurrentTime();a.mutedAutoplay=!1;a.endSeconds=NaN;a.be();this.u.loadVideoById(a.videoId,b);g.kL(this.u,this.element);this.hide()};g.r(O4,g.Q);O4.prototype.w=function(){var a=g.MK(this.u),b=this.u.getVideoData(),c=this.u.N();b=g.R(c.experiments,"embeds_enable_muted_autoplay")&&b.mutedAutoplay;c=!c.B;!b||g.V(a,2)?this.hide():this.Fa()||(g.Q.prototype.show.call(this),c||(this.A.style.display="none"),this.B.start(),g.lL(this.u,this.element,this.Fa()))};
-O4.prototype.C=function(){g.K(this.element,"ytp-muted-autoplay-hide-watermark",!0)};
-O4.prototype.F=function(){var a=this.u.getVideoData(),b=this.u.getCurrentTime();a.mutedAutoplay=!1;a.endSeconds=NaN;a.be();this.u.loadVideoById(a.videoId,b);g.kL(this.u,this.element)};g.r(P4,g.Q);g.h=P4.prototype;g.h.hide=function(){g.Ln(this.u.getRootNode(),"ytp-expand-pause-overlay");g.Q.prototype.hide.call(this)};
-g.h.HK=function(){this.aa=!0;g.Ln(this.u.getRootNode(),"ytp-expand-pause-overlay");this.H.focus()};
-g.h.IK=function(){this.aa=!1;g.I(this.u.getRootNode(),"ytp-expand-pause-overlay")};
-g.h.TH=function(){rua(this,this.B-this.F)};
-g.h.UH=function(){rua(this,this.B+this.F)};
-g.h.VH=function(a,b){if(1E3>(0,g.N)()-this.na)g.Op(b),document.activeElement.blur();else{var c=this.w[a],d=c.Lb;if(g.zM(b,this.u,this.da,d||void 0)){var e=c.va().videoId;c=c.getPlaylistId();g.oT(this.u.app,e,d,c,void 0,void 0)}}};
-g.h.WH=function(){this.it(g.NK(this.u).getPlayerSize())};
-g.h.XH=function(a){if(!(g.V(a.state,1)||g.V(a.state,16)||g.V(a.state,32))){var b=!g.R(this.u.N().experiments,"embeds_disable_pauseoverlay_on_autoplay_blocked_killswitch")&&g.V(a.state,2048);!g.V(a.state,4)||g.V(a.state,2)||b?this.T.hide():this.w.length&&(this.aa||(g.I(this.u.getRootNode(),"ytp-expand-pause-overlay"),Q4(this)),this.T.show(),this.na=(0,g.N)())}};
-g.h.it=function(a){var b=16/9,c=this.ca.qc();a=a.width-(c?112:58);c=Math.ceil(a/(c?320:192));c=(a-8*c)/c;b=Math.floor(c/b);for(var d=0;d<this.C.length;d++){var e=this.C[d].o["ytp-suggestion-image"];e.style.width=c+"px";e.style.height=b+"px"}this.G.element.style.height=b+"px";this.Z=c;this.ha=b;this.F=a;this.B=0;this.ez(0);Q4(this)};
-g.h.dz=function(){var a=this,b=this.u.N(),c=this.u.getVideoData();this.da=c.Ac?!1:b.u;if(c.suggestions){var d=(0,g.we)(c.suggestions,function(e){return e&&!e.list});
-this.w=(0,g.Ec)(d,function(e){e=g.VN(b,e);g.C(a,e);return e})}else this.w.length=0;
-sua(this);c.Ac?this.R.update({title:g.tM("More videos from $DNI_RELATED_CHANNEL",{DNI_RELATED_CHANNEL:c.author})}):this.R.update({title:"More videos"})};
-g.h.ez=function(a){this.G.element.scrollLeft=-a};g.r(R4,g.Q);g.h=R4.prototype;g.h.jt=function(){var a=this.u.getVideoData(),b=this.u.N();a=a.mutedAutoplay||b.sa&&!g.V(this.w,2);g.Ft(this,a);g.lL(this.u,this.element,a)};
-g.h.cI=function(a){a=a.state;this.w!=a&&(this.w=a);this.jt()};
-g.h.iz=function(){if(this.u.getVideoData().videoId){var a=this.u.getVideoUrl(!0,!1,!1,!0);this.ka("url",a);this.A||(this.A=this.ia("click",this.bI))}else this.A&&(this.ka("url",null),this.Va(this.A),this.A=null)};
-g.h.bI=function(a){var b=this.u.getVideoUrl(!g.LL(a),!1,!0,!0);g.AM(b,this.u,a);g.kL(this.u,this.element)};
-g.h.hz=function(a){var b=this.u.N(),c=480>a.width;if(c&&!this.B||!c&&this.B)a=480>a.width&&b.B?{D:"div",Y:["ytp-icon","ytp-icon-watermark-small"]}:F4(),a=new g.Q(a),b=this.o["ytp-watermark"],g.K(b,"ytp-watermark-small",c),g.te(b),a.ba(b),this.B=c};g.r(S4,g.sL);g.h=S4.prototype;g.h.Mf=function(){return!1};
-g.h.create=function(){var a=this.o.N(),b=g.gC(this.o);a.Ha&&(this.G=new P4(this.o,b),g.C(this,this.G),g.fL(this.o,this.G.element,4));g.R(a.experiments,"embeds_enable_muted_autoplay")&&(this.w=new O4(this.o),g.C(this,this.w),g.fL(this.o,this.w.element,4),this.F=new N4(this.o),g.C(this,this.F),g.fL(this.o,this.F.element,4));if(a.sa||this.w)this.A=new R4(this.o),g.C(this,this.A),g.fL(this.o,this.A.element,7);g.R(a.experiments,"embeds_impression_link")&&(this.C=new M4(this.o,b),g.C(this,this.C),g.fL(this.o,
-this.C.element,7));this.B.L(this.o,"appresize",this.QH);this.B.L(this.o,"presentingplayerstatechange",this.PH);this.B.L(this.o,"videodatachange",this.bP);this.B.L(this.o,"onMutedAutoplayStarts",this.PL);tua(this,g.MK(this.o));g.cL(this.player,"embed")};
-g.h.QH=function(){var a=g.NK(this.o).getPlayerSize();this.u&&this.u.u(a)};
-g.h.PH=function(a){tua(this,a.state)};
-g.h.PL=function(){this.o.getVideoData().mutedAutoplay&&this.w&&this.A&&this.A.ba(this.w.bottomButtons,0)};
-g.h.bP=function(){var a=this.o.getVideoData();this.A&&this.w&&!a.mutedAutoplay&&g.ze(this.w.element,this.A.element)&&g.fL(this.o,this.A.element,7)};g.HL.embed=S4;})(_yt_player);
