--- instructions/YOLO_ Real-Time Object Detection_files/remote.js
@@ -1,534 +1,0 @@
-(function(g){var window=this;var ova=function(a,b){return g.Jb(a,b)},Y5=function(a,b,c){a.w.set(b,c)},Z5=function(a){Y5(a,"zx",Math.floor(2147483648*Math.random()).toString(36)+Math.abs(Math.floor(2147483648*Math.random())^(0,g.G)()).toString(36));
-return a},$5=function(a,b,c){Array.isArray(c)||(c=[String(c)]);
-g.an(a.w,b,c)},pva=function(a,b){var c=[];
-g.jk(b,function(d){try{var e=g.fo.prototype.u.call(this,d,!0)}catch(f){if("Storage: Invalid value was encountered"==f)return;throw f;}void 0===e?c.push(d):g.eo(e)&&c.push(d)},a);
-return c},qva=function(a,b){var c=pva(a,b);
-(0,g.y)(c,function(d){g.fo.prototype.remove.call(this,d)},a)},rva=function(a){if(a.Tc){if(a.Tc.locationOverrideToken)return{locationOverrideToken:a.Tc.locationOverrideToken};
-if(null!=a.Tc.latitudeE7&&null!=a.Tc.longitudeE7)return{latitudeE7:a.Tc.latitudeE7,longitudeE7:a.Tc.longitudeE7}}return null},sva=function(a,b){g.bb(a,b)||a.push(b)},a6=function(a){var b=0,c;
-for(c in a)b++;return b},tva=function(a,b){var c=b instanceof g.uc?b:g.yc(b,/^data:image\//i.test(b));
-a.src=g.vc(c)},b6=function(){},uva=function(a){try{return g.v.JSON.parse(a)}catch(b){}a=String(a);
-if(/^\s*$/.test(a)?0:/^[\],:{}\s\u2028\u2029]*$/.test(a.replace(/\\["\\\/bfnrtu]/g,"@").replace(/(?:"[^"\\\n\r\u2028\u2029\x00-\x08\x0a-\x1f]*"|true|false|null|-?\d+(?:\.\d*)?(?:[eE][+\-]?\d+)?)[\s\u2028\u2029]*(?=:|,|]|}|$)/g,"]").replace(/(?:^|:|,)(?:[\s\u2028\u2029]*\[)+/g,"")))try{return eval("("+a+")")}catch(b){}throw Error("Invalid JSON string: "+a);},vva=function(a){if(a.Gd&&"function"==typeof a.Gd)return a.Gd();
-if("string"===typeof a)return a.split("");if(g.Na(a)){for(var b=[],c=a.length,d=0;d<c;d++)b.push(a[d]);return b}return g.Fb(a)},wva=function(a,b){if(a.forEach&&"function"==typeof a.forEach)a.forEach(b,void 0);
-else if(g.Na(a)||"string"===typeof a)(0,g.y)(a,b,void 0);else{if(a.xe&&"function"==typeof a.xe)var c=a.xe();else if(a.Gd&&"function"==typeof a.Gd)c=void 0;else if(g.Na(a)||"string"===typeof a){c=[];for(var d=a.length,e=0;e<d;e++)c.push(e)}else c=g.Gb(a);d=vva(a);e=d.length;for(var f=0;f<e;f++)b.call(void 0,d[f],c&&c[f],a)}},xva=function(a,b,c,d){var e=new g.Pm(null,void 0);
-a&&g.Qm(e,a);b&&g.Rm(e,b);c&&g.Sm(e,c);d&&(e.u=d);return e},c6=function(a,b){g.Qo[a]=!0;
-var c=g.Oo();c&&c.publish.apply(c,arguments);g.Qo[a]=!1},d6=function(a){this.app=this.name=this.id="";
-this.type="REMOTE_CONTROL";this.obfuscatedGaiaId=this.avatar=this.username="";this.w=!1;this.capabilities=new Set;this.experiments=new Set;this.theme="u";new g.Mm;this.o=this.u="";a&&(this.id=a.id||a.name,this.name=a.name,this.app=a.app,this.type=a.type||"REMOTE_CONTROL",this.username=a.user||"",this.avatar=a.userAvatarUri||"",this.obfuscatedGaiaId=a.obfuscatedGaiaId||"",this.theme=a.theme||"u",yva(this,a.capabilities||""),zva(this,a.experiments||""),this.u=a.remoteControllerUrl||"",this.o=a.localChannelEncryptionKey||
-"")},yva=function(a,b){a.capabilities.clear();
-(0,g.we)(b.split(","),g.Sa(ova,Ava)).forEach(function(c){a.capabilities.add(c)})},zva=function(a,b){a.experiments.clear();
-b.split(",").forEach(function(c){a.experiments.add(c)})},e6=function(a){a=a||{};
-this.name=a.name||"";this.id=a.id||a.screenId||"";this.token=a.token||a.loungeToken||"";this.uuid=a.uuid||a.dialId||""},f6=function(a,b){return!!b&&(a.id==b||a.uuid==b)},Bva=function(a){return{name:a.name,
-screenId:a.id,loungeToken:a.token,dialId:a.uuid}},Cva=function(a){return new e6(a)},Dva=function(a){return g.Ma(a)?(0,g.Ec)(a,Cva):[]},g6=function(a){return a?'{name:"'+a.name+'",id:'+a.id.substr(0,6)+"..,token:"+(a.token?".."+a.token.slice(-6):"-")+",uuid:"+(a.uuid?".."+a.uuid.slice(-6):"-")+"}":"null"},h6=function(a){return g.Ma(a)?"["+(0,g.Ec)(a,g6).join(",")+"]":"null"},i6=function(){return"xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g,function(a){var b=16*Math.random()|0;
-return("x"==a?b:b&3|8).toString(16)})},Eva=function(a){return(0,g.Ec)(a,function(b){return{key:b.id,
-name:b.name}})},j6=function(a,b){return g.Za(a,function(c){return c||b?!c!=!b?!1:c.id==b.id:!0})},k6=function(a,b){return g.Za(a,function(c){return f6(c,b)})},l6=function(a){try{var b=(0,g.Fs)(),c=(0,g.Es)();
-b&&b.remove(a);c&&c.remove(a)}catch(d){}},Fva=function(){var a=(0,g.Es)();
-a&&qva(a,a.o.Tf(!0))},m6=function(){var a=g.Hs("yt-remote-connected-devices")||[];
-g.sb(a);return a},Gva=function(a){if(g.cb(a))return[];
-var b=a[0].indexOf("#"),c=-1==b?a[0]:a[0].substring(0,b);return(0,g.Ec)(a,function(d,e){return 0==e?d:d.substring(c.length)})},Hva=function(a){g.Gs("yt-remote-connected-devices",a,86400)},o6=function(){if(n6)return n6;
-var a=g.Hs("yt-remote-device-id");a||(a=i6(),g.Gs("yt-remote-device-id",a,31536E3));for(var b=m6(),c=1,d=a;g.bb(b,d);)c++,d=a+"#"+c;return n6=d},p6=function(){var a=m6(),b=o6();
-g.bb(a,b);g.Is()&&g.ub(a,b);a=Gva(a);if(g.cb(a))try{g.Sq("remote_sid")}catch(c){}else try{g.Qq("remote_sid",a.join(","),-1)}catch(c){}},Iva=function(){return g.Hs("yt-remote-session-browser-channel")},Jva=function(){return g.Hs("yt-remote-local-screens")||[]},Kva=function(){g.Gs("yt-remote-lounge-token-expiration",!0,86400)},Lva=function(a){5<a.length&&(a=a.slice(a.length-5));
-var b=(0,g.Ec)(Jva(),function(d){return d.loungeToken}),c=(0,g.Ec)(a,function(d){return d.loungeToken});
-(0,g.ui)(c,function(d){return!g.bb(b,d)})&&Kva();
-g.Gs("yt-remote-local-screens",a,31536E3)},Mva=function(a,b){g.Gs("yt-remote-session-browser-channel",a);
-g.Gs("yt-remote-session-screen-id",b);var c=m6(),d=o6();g.bb(c,d)||c.push(d);Hva(c);p6()},q6=function(a){a||(l6("yt-remote-session-screen-id"),l6("yt-remote-session-video-id"));
-p6();a=m6();g.fb(a,o6());Hva(a)},Nva=function(){if(!r6){var a=g.po();
-a&&(r6=new g.Zn(a))}return r6?!!r6.get("yt-remote-use-staging-server"):!1},Ova=function(a){return!!document.currentScript&&(-1!=document.currentScript.src.indexOf("?"+a)||-1!=document.currentScript.src.indexOf("&"+a))},Pva=function(){return"function"==typeof window.__onGCastApiAvailable?window.__onGCastApiAvailable:null},s6=function(a){a.length?Qva(a.shift(),function(){s6(a)}):t6()},Rva=function(a){return"chrome-extension://"+a+"/cast_sender.js"},Qva=function(a,b,c){var d=document.createElement("script");
-d.onerror=b;c&&(d.onload=c);d.src=a;(document.head||document.documentElement).appendChild(d)},t6=function(){var a=Pva();
-a&&a(!1,"No cast extension found")},Tva=function(){if(Sva){var a=2,b=Pva(),c=function(){a--;
-0==a&&b&&b(!0)};
-window.__onGCastApiAvailable=c;Qva("//www.gstatic.com/cast/sdk/libs/sender/1.0/cast_framework.js",t6,c)}},Uva=function(){Tva();
-var a=window.navigator.userAgent.match(/Chrome\/([0-9]+)/);s6(["//www.gstatic.com/eureka/clank/"+(a?parseInt(a[1],10):0)+"/cast_sender.js","//www.gstatic.com/eureka/clank/cast_sender.js"])},u6=function(a,b,c){g.A.call(this);
-this.C=null!=c?(0,g.x)(a,c):a;this.Cd=b;this.B=(0,g.x)(this.oF,this);this.o=!1;this.u=0;this.w=this.Ea=null;this.A=[]},v6=function(a,b,c){g.A.call(this);
-this.w=null!=c?(0,g.x)(a,c):a;this.Cd=b;this.u=(0,g.x)(this.pF,this);this.o=[]},w6=function(a){a.Ea=g.Vf(a.u,a.Cd);
-a.w.apply(null,a.o)},x6=function(a){if(g.v.JSON)try{return g.v.JSON.parse(a)}catch(b){}return uva(a)},y6=function(){},z6=function(){},Vva=function(){},Xva=function(a){return(a=Wva(a))?new ActiveXObject(a):new XMLHttpRequest},Wva=function(a){if(!a.u&&"undefined"==typeof XMLHttpRequest&&"undefined"!=typeof ActiveXObject){for(var b=["MSXML2.XMLHTTP.6.0",
-"MSXML2.XMLHTTP.3.0","MSXML2.XMLHTTP","Microsoft.XMLHTTP"],c=0;c<b.length;c++){var d=b[c];try{return new ActiveXObject(d),a.u=d}catch(e){}}throw Error("Could not create ActiveXObject. ActiveX might be disabled, or MSXML might not be installed");}return a.u},A6=function(a,b,c,d){this.o=a;
-this.w=b;this.G=c;this.F=d||1;this.B=45E3;this.A=new g.I1(this);this.u=new g.Uf;this.u.setInterval(250)},Zva=function(a,b,c){a.rj=1;
-a.Pg=Z5(b.clone());a.Li=c;a.C=!0;Yva(a,null)},B6=function(a,b,c,d,e){a.rj=1;
-a.Pg=Z5(b.clone());a.Li=null;a.C=c;e&&(a.ZB=!1);Yva(a,d)},Yva=function(a,b){a.ak=(0,g.G)();
-C6(a);a.Eh=a.Pg.clone();$5(a.Eh,"t",a.F);a.Lm=0;a.Lc=a.o.kr(a.o.Bm()?b:null);0<a.Vr&&(a.zp=new v6((0,g.x)(a.WC,a,a.Lc),a.Vr));a.A.ia(a.Lc,"readystatechange",a.cO);var c=a.ni?g.Pb(a.ni):{};a.Li?(a.fq="POST",c["Content-Type"]="application/x-www-form-urlencoded",a.Lc.send(a.Eh,a.fq,a.Li,c)):(a.fq="GET",a.ZB&&!g.le&&(c.Connection="close"),a.Lc.send(a.Eh,a.fq,null,c));a.o.af(1)},bwa=function(a,b,c){for(var d=!0;!a.Wh&&a.Lm<c.length;){var e=$va(a,c);
-if(e==D6){4==b&&(a.fh=4,E6(15),d=!1);break}else if(e==awa){a.fh=4;E6(16);d=!1;break}else F6(a,e)}4==b&&0==c.length&&(a.fh=1,E6(17),d=!1);a.bf=a.bf&&d;d||(G6(a),H6(a))},$va=function(a,b){var c=a.Lm,d=b.indexOf("\n",c);
-if(-1==d)return D6;c=Number(b.substring(c,d));if(isNaN(c))return awa;d+=1;if(d+c>b.length)return D6;var e=b.substr(d,c);a.Lm=d+c;return e},dwa=function(a,b){a.ak=(0,g.G)();
-C6(a);var c=b?window.location.hostname:"";a.Eh=a.Pg.clone();Y5(a.Eh,"DOMAIN",c);Y5(a.Eh,"t",a.F);try{a.gf=new ActiveXObject("htmlfile")}catch(n){G6(a);a.fh=7;E6(22);H6(a);return}var d="<html><body>";if(b){for(var e="",f=0;f<c.length;f++){var k=c.charAt(f);if("<"==k)e+="\\x3c";else if(">"==k)e+="\\x3e";else{var l=k;if(l in I6)k=I6[l];else if(l in cwa)k=I6[l]=cwa[l];else{var m=l.charCodeAt(0);if(31<m&&127>m)k=l;else{if(256>m){if(k="\\x",16>m||256<m)k+="0"}else k="\\u",4096>m&&(k+="0");k+=m.toString(16).toUpperCase()}k=
-I6[l]=k}e+=k}}d+='<script>document.domain="'+e+'"\x3c/script>'}c=g.Yc(g.Yb("b/12014412"),d+"</body></html>");a.gf.open();a.gf.write(g.Tc(c));a.gf.close();a.gf.parentWindow.m=(0,g.x)(a.xN,a);a.gf.parentWindow.d=(0,g.x)(a.hB,a,!0);a.gf.parentWindow.rpcClose=(0,g.x)(a.hB,a,!1);c=a.gf.createElement("DIV");a.gf.parentWindow.document.body.appendChild(c);d=g.xc(a.Eh.toString());d=g.fd(g.vc(d));d=g.Yc(g.Yb("b/12014412"),'<iframe src="'+d+'"></iframe>');g.Zc(c,d);a.o.af(1)},C6=function(a){a.Zu=(0,g.G)()+a.B;
-ewa(a,a.B)},ewa=function(a,b){if(null!=a.rk)throw Error("WatchDog timer not null");
-a.rk=J6((0,g.x)(a.FN,a),b)},K6=function(a){a.rk&&(g.v.clearTimeout(a.rk),a.rk=null)},H6=function(a){a.o.Dx()||a.Wh||a.o.eo(a)},G6=function(a){K6(a);
-g.Je(a.zp);a.zp=null;a.u.stop();g.qqa(a.A);if(a.Lc){var b=a.Lc;a.Lc=null;b.abort();b.dispose()}a.gf&&(a.gf=null)},F6=function(a,b){try{a.o.ZA(a,b),a.o.af(4)}catch(c){}},gwa=function(a,b,c,d,e){if(0==d)c(!1);
-else{var f=e||0;d--;fwa(a,b,function(k){k?c(!0):g.v.setTimeout(function(){gwa(a,b,c,d,f)},f)})}},fwa=function(a,b,c){var d=new Image;
-d.onload=function(){try{L6(d),c(!0)}catch(e){}};
-d.onerror=function(){try{L6(d),c(!1)}catch(e){}};
-d.onabort=function(){try{L6(d),c(!1)}catch(e){}};
-d.ontimeout=function(){try{L6(d),c(!1)}catch(e){}};
-g.v.setTimeout(function(){if(d.ontimeout)d.ontimeout()},b);
-tva(d,a)},L6=function(a){a.onload=null;
-a.onerror=null;a.onabort=null;a.ontimeout=null},hwa=function(a){this.o=a;
-this.u=new y6},iwa=function(a){var b=M6(a.o,a.Ck,"/mail/images/cleardot.gif");
-Z5(b);gwa(b.toString(),5E3,(0,g.x)(a.jE,a),3,2E3);a.af(1)},O6=function(a){var b=a.o.H;
-if(null!=b)E6(5),b?(E6(11),N6(a.o,a,!1)):(E6(12),N6(a.o,a,!0));else if(a.Xd=new A6(a,void 0,void 0,void 0),a.Xd.ni=a.Tr,b=a.o,b=M6(b,b.Bm()?a.il:null,a.Ur),E6(5),!g.je||g.Nd(10))$5(b,"TYPE","xmlhttp"),B6(a.Xd,b,!1,a.il,!1);else{$5(b,"TYPE","html");var c=a.Xd;a=!!a.il;c.rj=3;c.Pg=Z5(b.clone());dwa(c,a)}},P6=function(a){g.jf.call(this);
-this.headers=new g.Mm;this.R=a||null;this.w=!1;this.V=this.o=null;this.da=this.H="";this.C=0;this.A="";this.B=this.aa=this.G=this.T=!1;this.F=0;this.J=null;this.ea="";this.M=this.Z=!1},jwa=function(a){return g.je&&g.Md(9)&&"number"===typeof a.timeout&&void 0!==a.ontimeout},kwa=function(a){return"content-type"==a.toLowerCase()},mwa=function(a,b){a.w=!1;
-a.o&&(a.B=!0,a.o.abort(),a.B=!1);a.A=b;a.C=5;lwa(a);Q6(a)},lwa=function(a){a.T||(a.T=!0,a.dispatchEvent("complete"),a.dispatchEvent("error"))},owa=function(a){if(a.w&&"undefined"!=typeof g.N1)if(a.V[1]&&4==R6(a)&&2==a.getStatus())S6(a,"Local request error detected and ignored");
-else if(a.G&&4==R6(a))g.Vf(a.VA,0,a);else if(a.dispatchEvent("readystatechange"),4==R6(a)){S6(a,"Request complete");a.w=!1;try{var b=a.getStatus();a:switch(b){case 200:case 201:case 202:case 204:case 206:case 304:case 1223:var c=!0;break a;default:c=!1}var d;if(!(d=c)){var e;if(e=0===b){var f=g.pd(1,String(a.H));if(!f&&g.v.self&&g.v.self.location){var k=g.v.self.location.protocol;f=k.substr(0,k.length-1)}e=!nwa.test(f?f.toLowerCase():"")}d=e}if(d)a.dispatchEvent("complete"),a.dispatchEvent("success");
-else{a.C=6;try{var l=2<R6(a)?a.o.statusText:""}catch(m){l=""}a.A=l+" ["+a.getStatus()+"]";lwa(a)}}finally{Q6(a)}}},Q6=function(a,b){if(a.o){pwa(a);
-var c=a.o,d=a.V[0]?g.Ja:null;a.o=null;a.V=null;b||a.dispatchEvent("ready");try{c.onreadystatechange=d}catch(e){}}},pwa=function(a){a.o&&a.M&&(a.o.ontimeout=null);
-a.J&&(g.v.clearTimeout(a.J),a.J=null)},R6=function(a){return a.o?a.o.readyState:0},T6=function(a){try{return a.o?a.o.responseText:""}catch(b){return""}},S6=function(a,b){return b+" ["+a.da+" "+a.H+" "+a.getStatus()+"]"},U6=function(a,b,c){this.o=1;
-this.u=[];this.A=[];this.B=new y6;this.G=a||null;this.H=null!=b?b:null;this.C=c||!1},qwa=function(a,b){this.o=a;
-this.map=b;this.context=null},rwa=function(a){g.Le.call(this,"statevent",a)},swa=function(a,b){g.Le.call(this,"timingevent",a);
-this.size=b},twa=function(a){g.Le.call(this,"serverreachability",a)},wwa=function(a){uwa(a);
-if(3==a.o){var b=a.Xl++,c=a.Qn.clone();Y5(c,"SID",a.w);Y5(c,"RID",b);Y5(c,"TYPE","terminate");V6(a,c);b=new A6(a,a.w,b,void 0);b.rj=2;b.Pg=Z5(c.clone());tva(new Image,b.Pg.toString());b.ak=(0,g.G)();C6(b)}vwa(a)},xwa=function(a){a.IE(1,0);
-a.Qn=M6(a,null,a.Sr);W6(a)},uwa=function(a){a.Tg&&(a.Tg.abort(),a.Tg=null);
-a.pc&&(a.pc.cancel(),a.pc=null);a.Vf&&(g.v.clearTimeout(a.Vf),a.Vf=null);X6(a);a.Nd&&(a.Nd.cancel(),a.Nd=null);a.Yg&&(g.v.clearTimeout(a.Yg),a.Yg=null)},ywa=function(a,b){if(0==a.o)throw Error("Invalid operation: sending map when state is closed");
-a.u.push(new qwa(a.RJ++,b));2!=a.o&&3!=a.o||W6(a)},W6=function(a){a.Nd||a.Yg||(a.Yg=J6((0,g.x)(a.gB,a),0),a.kj=0)},Awa=function(a,b){if(1==a.o){if(!b){a.Xl=Math.floor(1E5*Math.random());
-var c=a.Xl++,d=new A6(a,"",c,void 0);d.ni=null;var e=Y6(a),f=a.Qn.clone();Y5(f,"RID",c);Y5(f,"CVER","1");V6(a,f);Zva(d,f,e);a.Nd=d;a.o=2}}else 3==a.o&&(b?zwa(a,b):0==a.u.length||a.Nd||zwa(a))},zwa=function(a,b){if(b)if(6<a.Xh){a.u=a.A.concat(a.u);
-a.A.length=0;var c=a.Xl-1;var d=Y6(a)}else c=b.G,d=b.Li;else c=a.Xl++,d=Y6(a);var e=a.Qn.clone();Y5(e,"SID",a.w);Y5(e,"RID",c);Y5(e,"AID",a.Ej);V6(a,e);c=new A6(a,a.w,c,a.kj+1);c.ni=null;c.setTimeout(1E4+Math.round(1E4*Math.random()));a.Nd=c;Zva(c,e,d)},V6=function(a,b){if(a.yd){var c=a.yd.ax();
-c&&g.zb(c,function(d,e){Y5(b,e,d)})}},Y6=function(a){var b=Math.min(a.u.length,1E3),c=["count="+b];
-if(6<a.Xh&&0<b){var d=a.u[0].o;c.push("ofs="+d)}else d=0;for(var e=0;e<b;e++){var f=a.u[e].o,k=a.u[e].map;f=6>=a.Xh?e:f-d;try{g.zb(k,function(l,m){c.push("req"+f+"_"+m+"="+encodeURIComponent(l))})}catch(l){c.push("req"+f+"_type="+encodeURIComponent("_badmap"))}}a.A=a.A.concat(a.u.splice(0,b));
-return c.join("&")},Bwa=function(a){a.pc||a.Vf||(a.F=1,a.Vf=J6((0,g.x)(a.fB,a),0),a.cj=0)},Z6=function(a){if(a.pc||a.Vf||3<=a.cj)return!1;
-a.F++;a.Vf=J6((0,g.x)(a.fB,a),Cwa(a,a.cj));a.cj++;return!0},N6=function(a,b,c){a.cq=c;
-a.sf=b.lg;a.C||xwa(a)},X6=function(a){null!=a.di&&(g.v.clearTimeout(a.di),a.di=null)},Cwa=function(a,b){var c=5E3+Math.floor(1E4*Math.random());
-a.isActive()||(c*=2);return c*b},$6=function(a,b){if(2==b||9==b){var c=null;
-a.yd&&(c=null);var d=(0,g.x)(a.OO,a);c||(c=new g.Pm("//www.google.com/images/cleardot.gif"),Z5(c));fwa(c.toString(),1E4,d)}else E6(2);Dwa(a,b)},Dwa=function(a,b){a.o=0;
-a.yd&&a.yd.qw(b);vwa(a);uwa(a)},vwa=function(a){a.o=0;
-a.sf=-1;if(a.yd)if(0==a.A.length&&0==a.u.length)a.yd.Zq();else{g.jb(a.A);var b=g.jb(a.u);a.A.length=0;a.u.length=0;a.yd.Zq(b)}},M6=function(a,b,c){var d=g.Xm(c);
-if(""!=d.o)b&&g.Rm(d,b+"."+d.o),g.Sm(d,d.A);else{var e=window.location;d=xva(e.protocol,b?b+"."+e.hostname:e.hostname,+e.port,c)}a.Rk&&g.zb(a.Rk,function(f,k){Y5(d,k,f)});
-Y5(d,"VER",a.Xh);V6(a,d);return d},J6=function(a,b){if(!g.Oa(a))throw Error("Fn must not be null and must be a function");
-return g.v.setTimeout(function(){a()},b)},E6=function(a){a7.dispatchEvent(new rwa(a7,a))},Ewa=function(){},Fwa=function(){this.o=[];
-this.u=[]},Gwa=function(a,b){this.action=a;
-this.params=b||{}},b7=function(a,b){g.A.call(this);
-this.o=new g.H(this.oN,0,this);g.C(this,this.o);this.Cd=5E3;this.u=0;if(g.Oa(a))b&&(a=(0,g.x)(a,b));else if(a&&g.Oa(a.handleEvent))a=(0,g.x)(a.handleEvent,a);else throw Error("Invalid listener argument");this.w=a},c7=function(a,b,c){this.J=a;
-this.C=b;this.w=new g.Yn;this.u=new b7(this.qO,this);this.o=null;this.vb=!1;this.B=null;this.H="";this.G=this.A=0;this.F=[];this.M=c||!1},Hwa=function(a){return{firstTestResults:[""],
-secondTestResults:!a.o.cq,sessionId:a.o.w,arrayId:a.o.Ej}},Iwa=function(a,b){a.G=b||0;
-a.u.stop();a.o&&(3==a.o.o&&Awa(a.o),wwa(a.o));a.G=0},d7=function(a){return!!a.o&&3==a.o.o},Jwa=function(a,b){(a.C.loungeIdToken=b)||a.u.stop()},e7=function(a){this.port=this.domain="";
-this.o="/api/lounge";this.u=!0;a=a||document.location.href;var b=Number(g.pd(4,a))||"";b&&(this.port=":"+b);this.domain=g.qd(a)||"";a=g.Lc;0<=a.search("MSIE")&&(a=a.match(/MSIE ([\d.]+)/)[1],0>g.qc(a,"10.0")&&(this.u=!1))},f7=function(a,b){var c=a.o;
-a.u&&(c="https://"+a.domain+a.port+a.o);return g.Ad(c+b,{})},g7=function(a,b,c,d,e){a={format:"JSON",
-method:"POST",context:a,timeout:5E3,withCredentials:!1,onSuccess:g.Sa(a.A,d,!0),onError:g.Sa(a.w,e),Ud:g.Sa(a.B,e)};c&&(a.ub=c,a.headers={"Content-Type":"application/x-www-form-urlencoded"});return g.Iq(b,a)},Nwa=function(){var a=Kwa;
-Lwa();h7.push(a);Mwa(h7)},i7=function(a,b){Lwa();
-var c=h7,d=Owa(a,String(b));g.cb(c)?Pwa(d):(Mwa(c),(0,g.y)(c,function(e){e(d)}))},Lwa=function(){h7||(h7=g.w("yt.mdx.remote.debug.handlers_")||[],g.Ia("yt.mdx.remote.debug.handlers_",h7,void 0))},Pwa=function(a){var b=(j7+1)%50;
-j7=b;k7[b]=a;l7||(l7=49==b)},Mwa=function(a){var b=k7;
-if(b[0]){var c=j7,d=l7?c:-1;do{d=(d+1)%50;var e=b[d];(0,g.y)(a,function(f){f(e)})}while(d!=c);
-k7=Array(50);j7=-1;l7=!1}},Owa=function(a,b){var c=((0,g.G)()-Qwa)/1E3;
-c.toFixed&&(c=c.toFixed(3));var d=[];d.push("[",c+"s","] ");d.push("[","yt.mdx.remote","] ");d.push(a+": "+b,"\n");return d.join("")},m7=function(a){g.P.call(this);
-this.F=a;this.o=[]},Rwa=function(a,b){var c=a.get(b.uuid)||a.get(b.id);
-if(c){var d=c.name;c.id=b.id||c.id;c.name=b.name;c.token=b.token;c.uuid=b.uuid||c.uuid;return c.name!=d}a.o.push(b);return!0},Swa=function(a,b){var c=a.o.length!=b.length;
-a.o=(0,g.we)(a.o,function(f){return!!j6(b,f)});
-for(var d=0,e=b.length;d<e;d++)c=Rwa(a,b[d])||c;return c},Twa=function(a,b){var c=a.o.length;
-a.o=(0,g.we)(a.o,function(d){return!(d||b?!d!=!b?0:d.id==b.id:1)});
-return a.o.length<c},n7=function(a,b,c,d){g.P.call(this);
-this.C=a;this.A=b;this.B=c;this.w=d;this.u=0;this.o=null;this.Ea=NaN},p7=function(a){m7.call(this,"LocalScreenService");
-this.w=a;this.u=NaN;o7(this);this.info("Initializing with "+h6(this.o))},Uwa=function(a){if(a.o.length){var b=(0,g.Ec)(a.o,function(d){return d.id}),c=f7(a.w,"/pairing/get_lounge_token_batch");
-g7(a.w,c,{screen_ids:b.join(",")},(0,g.x)(a.zF,a),(0,g.x)(a.yF,a))}},o7=function(a){var b=Dva(Jva());
-b=(0,g.we)(b,function(c){return!c.uuid});
-return Swa(a,b)},q7=function(a,b){Lva((0,g.Ec)(a.o,Bva));
-b&&Kva()},s7=function(a,b){g.P.call(this);
-this.C=b;var c=g.Hs("yt-remote-online-screen-ids")||"";c=c?c.split(","):[];for(var d={},e=this.C(),f=0,k=e.length;f<k;++f){var l=e[f].id;d[l]=g.bb(c,l)}this.o=d;this.B=a;this.w=this.A=NaN;this.u=null;r7("Initialized with "+g.Lk(this.o))},Vwa=function(a,b,c){var d=f7(a.B,"/pairing/get_screen_availability");
-g7(a.B,d,{lounge_token:b.token},(0,g.x)(function(e){e=e.screens||[];for(var f=0,k=e.length;f<k;++f)if(e[f].loungeToken==b.token){c("online"==e[f].status);return}c(!1)},a),(0,g.x)(function(){c(!1)},a))},t7=function(a,b){a:if(a6(b)!=a6(a.o))var c=!1;
-else{c=g.Gb(b);for(var d=0,e=c.length;d<e;++d)if(!a.o[c[d]]){c=!1;break a}c=!0}c||(r7("Updated online screens: "+g.Lk(a.o)),a.o=b,a.P("screenChange"));Wwa(a)},u7=function(a){isNaN(a.w)||g.No(a.w);
-a.w=g.Lo((0,g.x)(a.qu,a),0<a.A&&a.A<(0,g.G)()?2E4:1E4)},r7=function(a){i7("OnlineScreenService",a)},Xwa=function(a){var b={};
-(0,g.y)(a.C(),function(c){c.token?b[c.token]=c.id:this.Sb("Requesting availability of screen w/o lounge token.")});
-return b},Wwa=function(a){a=g.Gb(g.Ab(a.o,function(b){return b}));
-g.sb(a);a.length?g.Gs("yt-remote-online-screen-ids",a.join(","),60):l6("yt-remote-online-screen-ids")},v7=function(a){m7.call(this,"ScreenService");
-this.C=a;this.u=this.w=null;this.A=[];this.B={};Ywa(this)},$wa=function(a,b,c,d,e,f){a.info("getAutomaticScreenByIds "+c+" / "+b);
-c||(c=a.B[b]);var k=a.le();if(k=(c?k6(k,c):null)||k6(k,b)){k.uuid=b;var l=w7(a,k);Vwa(a.u,l,function(m){e(m?l:null)})}else c?Zwa(a,c,(0,g.x)(function(m){var n=w7(this,new e6({name:d,
-screenId:c,loungeToken:m,dialId:b||""}));Vwa(this.u,n,function(p){e(p?n:null)})},a),f):e(null)},axa=function(a,b){for(var c=0,d=a.o.length;c<d;++c)if(a.o[c].name==b)return a.o[c];
-return null},Zwa=function(a,b,c,d){a.info("requestLoungeToken_ for "+b);
-var e={ub:{screen_ids:b},method:"POST",context:a,onSuccess:function(f,k){var l=k&&k.screens||[];l[0]&&l[0].screenId==b?c(l[0].loungeToken):d(Error("Missing lounge token in token response"))},
-onError:function(){d(Error("Request screen lounge token failed"))}};
-g.Iq(f7(a.C,"/pairing/get_lounge_token_batch"),e)},bxa=function(a){a.o=a.w.le();
-var b=a.B,c={},d;for(d in b)c[b[d]]=d;b=0;for(d=a.o.length;b<d;++b){var e=a.o[b];e.uuid=c[e.id]||""}a.info("Updated manual screens: "+h6(a.o))},Ywa=function(a){x7(a);
-a.w=new p7(a.C);a.w.subscribe("screenChange",(0,g.x)(a.IF,a));bxa(a);a.A=Dva(g.Hs("yt-remote-automatic-screen-cache")||[]);x7(a);a.info("Initializing automatic screens: "+h6(a.A));a.u=new s7(a.C,(0,g.x)(a.le,a,!0));a.u.subscribe("screenChange",(0,g.x)(function(){this.P("onlineScreenChange")},a))},w7=function(a,b){var c=a.get(b.id);
-c?(c.uuid=b.uuid,b=c):((c=k6(a.A,b.uuid))?(c.id=b.id,c.token=b.token,b=c):a.A.push(b),g.Gs("yt-remote-automatic-screen-cache",(0,g.Ec)(a.A,Bva)));x7(a);a.B[b.uuid]=b.id;g.Gs("yt-remote-device-id-map",a.B,31536E3);return b},x7=function(a){a.B=g.Hs("yt-remote-device-id-map")||{}},y7=function(a,b,c){g.P.call(this);
-this.V=c;this.H=a;this.u=b;this.w=null},z7=function(a,b){i7(a.V,b)},A7=function(a,b){y7.call(this,a,b,"CastSession");
-this.o=null;this.A=0;this.C=(0,g.x)(this.yP,this);this.B=(0,g.x)(this.NN,this);this.A=g.Lo((0,g.x)(function(){cxa(this,null)},this),12E4)},dxa=function(a){a.info("sendYoutubeMessage_: getMdxSessionStatus "+g.Lk(void 0));
-var b={type:"getMdxSessionStatus"};a.o?a.o.sendMessage("urn:x-cast:com.google.youtube.mdx",b,g.Ja,(0,g.x)(function(){z7(this,"Failed to send message: getMdxSessionStatus.")},a)):z7(a,"Sending yt message without session: "+g.Lk(b))},cxa=function(a,b){g.No(a.A);
-if(b){if(a.info("onConnectedScreenId_: Received screenId: "+b),!a.w||a.w.id!=b){var c=(0,g.x)(a.np,a),d=(0,g.x)(a.ne,a);a.wx(b,c,d,5)}}else a.ne(Error("Waiting for session status timed out."))},B7=function(a,b,c){y7.call(this,a,b,"DialSession");
-this.A=this.G=null;this.J="";this.R=c;this.B=null;this.F=g.Ja;this.C=NaN;this.M=(0,g.x)(this.BP,this);this.o=g.Ja},exa=function(a){a.o=a.H.cD(a.J,a.u.label,a.u.friendlyName,(0,g.x)(function(b){this.o=g.Ja;
-this.np(b)},a),(0,g.x)(function(b){this.o=g.Ja;
-this.ne(b)},a))},fxa=function(a){var b={};
-b.pairingCode=a.J;b.theme=a.R;if(a.B){var c=a.B.currentTime||0;b.v=a.B.videoId;b.t=c}Nva()&&(b.env_useStageMdx=1);return g.yd(b)},C7=function(a,b){y7.call(this,a,b,"ManualSession");
-this.o=g.Lo((0,g.x)(this.Fj,this,null),150)},D7=function(a,b,c,d){g.P.call(this);
-this.u=a;this.F=b||"233637DE";this.C=c||"cl";this.G=d||!1;this.o=null;this.B=!1;this.w=[];this.A=(0,g.x)(this.wM,this)},gxa=function(a,b){return b?g.Za(a.w,function(c){return f6(b,c.label)},a):null},E7=function(a){i7("Controller",a)},Kwa=function(a){window.chrome&&chrome.cast&&chrome.cast.logMessage&&chrome.cast.logMessage(a)},F7=function(a){return a.B||!!a.w.length||!!a.o},G7=function(a,b,c){b!=a.o&&(g.Je(a.o),(a.o=b)?(c?a.P("yt-remote-cast2-receiver-resumed",b.u):a.P("yt-remote-cast2-receiver-selected",
-b.u),b.subscribe("sessionScreen",(0,g.x)(a.eB,a,b)),b.w?a.P("yt-remote-cast2-session-change",b.w):c&&a.o.Fj(null)):a.P("yt-remote-cast2-session-change",null))},hxa=function(a){var b=a.u.bD(),c=a.o&&a.o.u;
-a=(0,g.Ec)(b,function(d){c&&f6(d,c.label)&&(c=null);var e=d.uuid?d.uuid:d.id,f=gxa(this,d);f?(f.label=e,f.friendlyName=d.name):(f=new chrome.cast.Receiver(e,d.name),f.receiverType=chrome.cast.ReceiverType.CUSTOM);return f},a);
-c&&(c.receiverType!=chrome.cast.ReceiverType.CUSTOM&&(c=new chrome.cast.Receiver(c.label,c.friendlyName),c.receiverType=chrome.cast.ReceiverType.CUSTOM),a.push(c));return a},nxa=function(a,b,c,d,e,f,k){ixa()?jxa(b,e,f,k)&&(I7(!0),window.chrome&&chrome.cast&&chrome.cast.isAvailable?kxa(a,c):(window.__onGCastApiAvailable=function(l,m){l?kxa(a,c):(J7("Failed to load cast API: "+m),K7(!1),I7(!1),l6("yt-remote-cast-available"),l6("yt-remote-cast-receiver"),lxa(),c(!1))},d?g.Xo("https://www.gstatic.com/cv/js/sender/v1/cast_sender.js"):
-0<=window.navigator.userAgent.indexOf("Android")&&0<=window.navigator.userAgent.indexOf("Chrome/")&&window.navigator.presentation?Uva():!window.chrome||!window.navigator.presentation||0<=window.navigator.userAgent.indexOf("Edge")?t6():(Tva(),s6(mxa.map(Rva))))):H7("Cannot initialize because not running Chrome")},lxa=function(){H7("dispose");
-var a=L7();a&&a.dispose();g.Ia("yt.mdx.remote.cloudview.instance_",null,void 0);oxa(!1);g.To(M7);M7.length=0},N7=function(){return!!g.Hs("yt-remote-cast-installed")},pxa=function(){var a=g.Hs("yt-remote-cast-receiver");
-return a?a.friendlyName:null},qxa=function(){H7("clearCurrentReceiver");
-l6("yt-remote-cast-receiver")},rxa=function(){return N7()?L7()?L7().getCastSession():(J7("getCastSelector: Cast is not initialized."),null):(J7("getCastSelector: Cast API is not installed!"),null)},P7=function(){N7()?L7()?O7()?(H7("Requesting cast selector."),L7().requestSession()):(H7("Wait for cast API to be ready to request the session."),M7.push(g.So("yt-remote-cast2-api-ready",P7))):J7("requestCastSelector: Cast is not initialized."):J7("requestCastSelector: Cast API is not installed!")},Q7=
-function(a,b){O7()?L7().setConnectedScreenStatus(a,b):J7("setConnectedScreenStatus called before ready.")},ixa=function(){var a=0<=g.Lc.search(/ (CrMo|Chrome|CriOS)\//);
-return g.ax||a},sxa=function(a,b){L7().init(a,b)},jxa=function(a,b,c,d){var e=!1;
-L7()||(a=new D7(a,b,c,d),a.subscribe("yt-remote-cast2-availability-change",function(f){g.Gs("yt-remote-cast-available",f);c6("yt-remote-cast2-availability-change",f)}),a.subscribe("yt-remote-cast2-receiver-selected",function(f){H7("onReceiverSelected: "+f.friendlyName);
-g.Gs("yt-remote-cast-receiver",f);c6("yt-remote-cast2-receiver-selected",f)}),a.subscribe("yt-remote-cast2-receiver-resumed",function(f){H7("onReceiverResumed: "+f.friendlyName);
-g.Gs("yt-remote-cast-receiver",f)}),a.subscribe("yt-remote-cast2-session-change",function(f){H7("onSessionChange: "+g6(f));
-f||l6("yt-remote-cast-receiver");c6("yt-remote-cast2-session-change",f)}),g.Ia("yt.mdx.remote.cloudview.instance_",a,void 0),e=!0);
-H7("cloudview.createSingleton_: "+e);return e},L7=function(){return g.w("yt.mdx.remote.cloudview.instance_")},kxa=function(a,b){K7(!0);
-I7(!1);sxa(a,function(c){c?(oxa(!0),g.Uo("yt-remote-cast2-api-ready")):(J7("Failed to initialize cast API."),K7(!1),l6("yt-remote-cast-available"),l6("yt-remote-cast-receiver"),lxa());b(c)})},H7=function(a){i7("cloudview",a)},J7=function(a){i7("cloudview",a)},K7=function(a){H7("setCastInstalled_ "+a);
-g.Gs("yt-remote-cast-installed",a)},O7=function(){return!!g.w("yt.mdx.remote.cloudview.apiReady_")},oxa=function(a){H7("setApiReady_ "+a);
-g.Ia("yt.mdx.remote.cloudview.apiReady_",a,void 0)},I7=function(a){g.Ia("yt.mdx.remote.cloudview.initializing_",a,void 0)},R7=function(a){this.index=-1;
-this.videoId=this.listId="";this.volume=this.playerState=-1;this.muted=!1;this.audioTrackId=null;this.C=this.F=0;this.o=null;this.hasNext=this.J=!1;this.H=this.G=this.u=this.B=0;this.A=NaN;this.w=!1;this.reset(a)},S7=function(a){a.audioTrackId=null;
-a.o=null;a.playerState=-1;a.J=!1;a.hasNext=!1;a.F=0;a.C=(0,g.G)();a.B=0;a.u=0;a.G=0;a.H=0;a.A=NaN;a.w=!1},T7=function(a){return a.Ya()?((0,g.G)()-a.C)/1E3:0},U7=function(a,b){a.F=b;
-a.C=(0,g.G)()},V7=function(a){switch(a.playerState){case 1:case 1081:return((0,g.G)()-a.C)/1E3+a.F;
-case -1E3:return 0}return a.F},W7=function(a,b,c){var d=a.videoId;
-a.videoId=b;a.index=c;b!=d&&S7(a)},X7=function(a){var b={};
-b.index=a.index;b.listId=a.listId;b.videoId=a.videoId;b.playerState=a.playerState;b.volume=a.volume;b.muted=a.muted;b.audioTrackId=a.audioTrackId;b.trackData=g.Qb(a.o);b.hasPrevious=a.J;b.hasNext=a.hasNext;b.playerTime=a.F;b.playerTimeAt=a.C;b.seekableStart=a.B;b.seekableEnd=a.u;b.duration=a.G;b.loadedTime=a.H;b.liveIngestionTime=a.A;return b},Z7=function(a,b){g.P.call(this);
-this.o=0;this.A=a;this.C=[];this.B=new Fwa;this.w=this.u=null;this.H=(0,g.x)(this.AK,this);this.F=(0,g.x)(this.Zl,this);this.G=(0,g.x)(this.zK,this);this.J=(0,g.x)(this.LK,this);var c=0;a?(c=a.getProxyState(),3!=c&&(a.subscribe("proxyStateChange",this.cv,this),txa(this))):c=3;0!=c&&(b?this.cv(c):g.Lo((0,g.x)(function(){this.cv(c)},this),0));
-var d=rxa();d&&Y7(this,d);this.subscribe("yt-remote-cast2-session-change",this.J)},$7=function(a){return new R7(a.A.getPlayerContextData())},txa=function(a){(0,g.y)("nowAutoplaying autoplayDismissed remotePlayerChange remoteQueueChange autoplayModeChange autoplayUpNext previousNextChange".split(" "),function(b){this.C.push(this.A.subscribe(b,g.Sa(this.tM,b),this))},a)},uxa=function(a){(0,g.y)(a.C,function(b){this.A.unsubscribeByKey(b)},a);
-a.C.length=0},a8=function(a,b){var c=a.B;
-50>c.o.length+c.u.length&&a.B.u.push(b)},c8=function(a,b,c){var d=$7(a);
-U7(d,c);-1E3!=d.playerState&&(d.playerState=b);b8(a,d)},d8=function(a,b,c){a.A.sendMessage(b,c)},b8=function(a,b){uxa(a);
-a.A.setPlayerContextData(X7(b));txa(a)},Y7=function(a,b){a.w&&(a.w.removeUpdateListener(a.H),a.w.removeMediaListener(a.F),a.Zl(null));
-a.w=b;a.w&&(i7("CP","Setting cast session: "+a.w.sessionId),a.w.addUpdateListener(a.H),a.w.addMediaListener(a.F),a.w.media.length&&a.Zl(a.w.media[0]))},vxa=function(a){var b=a.u.media,c=a.u.customData;
-if(b&&c){var d=$7(a);b.contentId!=d.videoId&&i7("CP","Cast changing video to: "+b.contentId);d.videoId=b.contentId;d.playerState=c.playerState;U7(d,a.u.getEstimatedTime());b8(a,d)}else i7("CP","No cast media video. Ignoring state update.")},e8=function(a,b,c){return(0,g.x)(function(d){this.Sb("Failed to "+b+" with cast v2 channel. Error code: "+d.code);
-d.code!=chrome.cast.ErrorCode.TIMEOUT&&(this.Sb("Retrying "+b+" using MDx browser channel."),d8(this,b,c))},a)},f8=function(a,b,c){g.P.call(this);
-this.B=NaN;this.M=!1;this.G=this.F=this.H=this.J=NaN;this.V=[];this.A=this.C=this.w=this.Sa=this.o=null;this.T=a;this.V.push(g.Ip(window,"beforeunload",(0,g.x)(this.rF,this)));this.u=[];this.Sa=new R7;this.R=b.id;this.o=wxa(this,c);this.o.subscribe("handlerOpened",this.EK,this);this.o.subscribe("handlerClosed",this.BK,this);this.o.subscribe("handlerError",this.CK,this);this.o.subscribe("handlerMessage",this.DK,this);Jwa(this.o,b.token);this.subscribe("remoteQueueChange",function(){var d=this.Sa.videoId;
-g.Is()&&g.Gs("yt-remote-session-video-id",d)},this)},g8=function(a){i7("conn",a)},wxa=function(a,b){return new c7(f7(a.T,"/bc"),b)},h8=function(a,b){a.P("proxyStateChange",b)},xxa=function(a){a.B=g.Lo((0,g.x)(function(){g8("Connecting timeout");
-this.ij(1)},a),2E4)},i8=function(a){g.No(a.B);
-a.B=NaN},j8=function(a){g.No(a.J);
-a.J=NaN},yxa=function(a){k8(a);
-a.H=g.Lo((0,g.x)(function(){l8(this,"getNowPlaying")},a),2E4)},k8=function(a){g.No(a.H);
-a.H=NaN},Axa=function(a,b){b&&(i8(a),j8(a));
-b==(d7(a.o)&&isNaN(a.B))?b&&(h8(a,1),l8(a,"getSubtitlesTrack")):b?(a.tx()&&a.Sa.reset(),h8(a,1),l8(a,"getNowPlaying"),zxa(a)):a.ij(1)},Bxa=function(a,b){var c=b.params.videoId;
-delete b.params.videoId;c==a.Sa.videoId&&(g.Mb(b.params)?a.Sa.o=null:a.Sa.o=b.params,a.P("remotePlayerChange"))},Cxa=function(a,b){var c=b.params.videoId||b.params.video_id,d=parseInt(b.params.currentIndex,10);
-a.Sa.listId=b.params.listId||a.Sa.listId;W7(a.Sa,c,d);a.P("remoteQueueChange")},Exa=function(a,b){b.params=b.params||{};
-Cxa(a,b);Dxa(a,b);a.P("autoplayDismissed")},Dxa=function(a,b){var c=parseInt(b.params.currentTime||b.params.current_time,10);
-U7(a.Sa,isNaN(c)?0:c);c=parseInt(b.params.state,10);c=isNaN(c)?-1:c;-1==c&&-1E3==a.Sa.playerState&&(c=-1E3);a.Sa.playerState=c;c=Number(b.params.loadedTime);a.Sa.H=isNaN(c)?0:c;c=Number(b.params.duration);a.Sa.G=isNaN(c)?0:c;c=a.Sa;var d=Number(b.params.liveIngestionTime);c.A=d;c.w=isNaN(d)?!1:!0;c=a.Sa;d=Number(b.params.seekableStartTime);var e=Number(b.params.seekableEndTime);c.B=isNaN(d)?0:d;c.u=isNaN(e)?0:e;1==a.Sa.playerState?yxa(a):k8(a);a.P("remotePlayerChange")},Fxa=function(a,b){if(-1E3!=
-a.Sa.playerState){var c=1085;
-switch(parseInt(b.params.adState,10)){case 1:c=1081;break;case 2:c=1084;break;case 0:c=1083}a.Sa.playerState=c;c=parseInt(b.params.currentTime,10);U7(a.Sa,isNaN(c)?0:c);a.P("remotePlayerChange")}},Gxa=function(a,b){var c="true"==b.params.muted;
-a.Sa.volume=parseInt(b.params.volume,10);a.Sa.muted=c;a.P("remotePlayerChange")},Hxa=function(a,b){a.C=b.params.videoId;
-a.P("nowAutoplaying",parseInt(b.params.timeout,10))},Ixa=function(a,b){var c="true"==b.params.hasNext;
-a.Sa.J="true"==b.params.hasPrevious;a.Sa.hasNext=c;a.P("previousNextChange")},zxa=function(a){g.No(a.G);
-a.G=g.Lo((0,g.x)(a.ij,a,1),864E5)},l8=function(a,b,c){c?g8("Sending: action="+b+", params="+g.Lk(c)):g8("Sending: action="+b);
-a.o.sendMessage(b,c)},m8=function(a){m7.call(this,"ScreenServiceProxy");
-this.Bd=a;this.u=[];this.u.push(this.Bd.$_s("screenChange",(0,g.x)(this.wP,this)));this.u.push(this.Bd.$_s("onlineScreenChange",(0,g.x)(this.WL,this)))},Nxa=function(a){var b={device:"Desktop",
-app:"youtube-desktop"};b=g.L("MDX_CONFIG")||b;Fva();p6();n8||(n8=new e7(b?b.loungeApiHost:void 0),Nva()&&(n8.o="/api/loungedev"));o8||(o8=g.w("yt.mdx.remote.deferredProxies_")||[],g.Ia("yt.mdx.remote.deferredProxies_",o8,void 0));Jxa();var c=p8();if(!c){var d=new v7(n8);g.Ia("yt.mdx.remote.screenService_",d,void 0);c=p8();var e=!1,f=void 0,k=void 0,l=!1;b&&(e=!!b.loadCastApiSetupScript,f=b.appId,k=b.theme,l=!!b.disableDial);nxa(a,d,function(m){m?q8()&&Q7(q8(),"YouTube TV"):d.subscribe("onlineScreenChange",
-function(){c6("yt-remote-receiver-availability-change")})},e,f,k,l)}b&&!g.w("yt.mdx.remote.initialized_")&&(g.Ia("yt.mdx.remote.initialized_",!0,void 0),r8("Initializing: "+g.Lk(b)),s8.push(g.So("yt-remote-cast2-availability-change",function(){c6("yt-remote-receiver-availability-change")})),s8.push(g.So("yt-remote-cast2-receiver-selected",function(){t8(null);
-c6("yt-remote-auto-connect","cast-selector-receiver")})),s8.push(g.So("yt-remote-cast2-receiver-resumed",function(){c6("yt-remote-receiver-resumed","cast-selector-receiver")})),s8.push(g.So("yt-remote-cast2-session-change",Kxa)),s8.push(g.So("yt-remote-connection-change",function(m){m?Q7(q8(),"YouTube TV"):u8()||(Q7(null,null),qxa())})),a=v8(),b.isAuto&&(a.id+="#dial"),g.Bo("desktop_enable_autoplay")&&(a.capabilities=["atp"]),a.name=b.device,a.app=b.app,(k=b.theme)&&(a.theme=k),r8(" -- with channel params: "+
-g.Lk(a)),Lxa(a),c.start(),q8()||Mxa())},Pxa=function(){var a=Oxa();
-N7()&&g.Hs("yt-remote-cast-available")&&a.push({key:"cast-selector-receiver",name:"Cast..."});return a},Oxa=function(){var a=p8().Bd.$_gos();
-var b=w8();b&&x8()&&(j6(a,b)||a.push(b));return Eva(a)},y8=function(){var a=Qxa();
-!a&&N7()&&pxa()&&(a={key:"cast-selector-receiver",name:pxa()});return a},Qxa=function(){var a=Oxa(),b=w8();
-b||(b=u8());return g.Za(a,function(c){return b&&f6(b,c.key)?!0:!1})},w8=function(){var a=q8();
-if(!a)return null;var b=p8().le();return k6(b,a)},Kxa=function(a){r8("remote.onCastSessionChange_: "+g6(a));
-if(a){var b=w8();b&&b.id==a.id?Q7(b.id,"YouTube TV"):(b&&z8(),A8(a,1))}else x8()&&z8()},z8=function(){O7()?L7().stopSession():J7("stopSession called before API ready.");
-var a=x8();a&&(a.disconnect(1),B8(null))},C8=function(){var a=x8();
-return!!a&&3!=a.getProxyState()},r8=function(a){i7("remote",a)},p8=function(){if(!D8){var a=g.w("yt.mdx.remote.screenService_");
-D8=a?new m8(a):null}return D8},q8=function(){return g.w("yt.mdx.remote.currentScreenId_")},Rxa=function(a){g.Ia("yt.mdx.remote.currentScreenId_",a,void 0)},Sxa=function(){return g.w("yt.mdx.remote.connectData_")},t8=function(a){g.Ia("yt.mdx.remote.connectData_",a,void 0)},x8=function(){return g.w("yt.mdx.remote.connection_")},B8=function(a){var b=x8();
-t8(null);a||Rxa("");g.Ia("yt.mdx.remote.connection_",a,void 0);o8&&((0,g.y)(o8,function(c){c(a)}),o8.length=0);
-b&&!a?c6("yt-remote-connection-change",!1):!b&&a&&c6("yt-remote-connection-change",!0)},u8=function(){var a=g.Is();
-if(!a)return null;var b=p8().le();return k6(b,a)},A8=function(a,b){q8();
-w8()&&w8();if(E8)F8=a;else{Rxa(a.id);var c=new f8(n8,a,v8());c.connect(b,Sxa());c.subscribe("beforeDisconnect",function(d){c6("yt-remote-before-disconnect",d)});
-c.subscribe("beforeDispose",function(){x8()&&(x8(),B8(null))});
-B8(c)}},Mxa=function(){var a=u8();
-a?(r8("Resume connection to: "+g6(a)),A8(a,0)):(q6(),qxa(),r8("Skipping connecting because no session screen found."))},Jxa=function(){var a=v8();
-if(g.Mb(a)){a=o6();var b=g.Hs("yt-remote-session-name")||"",c=g.Hs("yt-remote-session-app")||"";a={device:"REMOTE_CONTROL",id:a,name:b,app:c,mdxVersion:3};g.Ia("yt.mdx.remote.channelParams_",a,void 0)}},v8=function(){return g.w("yt.mdx.remote.channelParams_")||{}},Lxa=function(a){a?(g.Gs("yt-remote-session-app",a.app),g.Gs("yt-remote-session-name",a.name)):(l6("yt-remote-session-app"),l6("yt-remote-session-name"));
-g.Ia("yt.mdx.remote.channelParams_",a,void 0)},G8=function(a,b,c){g.A.call(this);
-this.B=a;this.u=b;this.w=new g.O(this);g.C(this,this.w);this.w.L(b,"onCaptionsTrackListChanged",this.uL);this.w.L(b,"captionschanged",this.yK);this.w.L(b,"captionssettingschanged",this.pC);this.w.L(b,"videoplayerreset",this.pp);this.w.L(b,"mdxautoplaycancel",this.iE);this.R=this.w.L(b,"onVolumeChange",this.DA);this.G=!1;this.o=c;c.ga();c.subscribe("proxyStateChange",this.YA,this);c.subscribe("remotePlayerChange",this.km,this);c.subscribe("remoteQueueChange",this.pp,this);c.subscribe("autoplayUpNext",
-this.oA,this);c.subscribe("previousNextChange",this.UA,this);c.subscribe("nowAutoplaying",this.PA,this);c.subscribe("autoplayDismissed",this.nA,this);this.suggestion=null;this.H=new g.MC(64);this.A=new g.H(this.nC,500,this);g.C(this,this.A);this.C=new g.H(this.oC,1E3,this);g.C(this,this.C);this.M=new u6(this.XO,0,this);g.C(this,this.M);this.F={};this.V=new g.H(this.OC,1E3,this);g.C(this,this.V);this.J=new v6(this.hI,1E3,this);g.C(this,this.J);this.T=g.Ja;this.pC();this.pp();this.km()},H8=function(a,
-b){var c=a.B,d=a.u.getVideoData().lengthSeconds;
-c.V=b||0;c.player.P("progresssync",b,d)},Txa=function(a){H8(a,0);
-a.A.stop();I8(a,new g.MC(64))},K8=function(a,b){if(J8(a)&&!a.G){var c=null;
-b&&(c={style:a.u.getSubtitlesUserSettings()},g.Sb(c,b));a.o.eC(a.u.getVideoData(1).videoId,c);a.F=$7(a.o).o}},L8=function(a,b){var c=a.u.getPlaylist();
-if(c){var d=c.index;var e=c.listId.toString()}c=a.u.getVideoData(1);a.o.playVideo(c.videoId,b,d,e,c.playerParams,c.uf,rva(c));I8(a,new g.MC(1))},Uxa=function(a,b){if(b){var c=a.u.getOption("captions","tracklist",{rT:1});
-c&&c.length?(a.u.setOption("captions","track",b),a.G=!1):(a.u.loadModule("captions"),a.G=!0)}else a.u.setOption("captions","track",{})},J8=function(a){return $7(a.o).videoId==a.u.getVideoData(1).videoId},I8=function(a,b){a.C.stop();
-var c=a.H;if(!g.SC(c,b)){var d=g.V(b,2);d!=g.V(a.H,2)&&g.QS(a.u.app,d);a.H=b;Vxa(a.B,c,b)}},M8=function(){g.Q.call(this,{D:"div",
-I:"ytp-mdx-popup-dialog",O:{role:"dialog"},K:[{D:"div",I:"ytp-mdx-popup-dialog-inner-content",K:[{D:"div",I:"ytp-mdx-popup-title",W:"You're signed out"},{D:"div",I:"ytp-mdx-popup-description",W:"Videos that you watch may be added to the TV's watch history and influence TV recommendations. To avoid this, cancel and sign in to YouTube on your computer."},{D:"div",I:"ytp-mdx-privacy-popup-buttons",K:[{D:"button",Y:["ytp-button","ytp-mdx-privacy-popup-cancel"],W:"Cancel"},{D:"button",Y:["ytp-button",
-"ytp-mdx-privacy-popup-confirm"],W:"Confirm"}]}]}]});this.u=new g.OL(this,250);g.C(this,this.u);this.C=this.o["ytp-mdx-privacy-popup-cancel"];this.L(this.C,"click",this.w);this.A=this.o["ytp-mdx-privacy-popup-confirm"];this.L(this.A,"click",this.B)},N8=function(a){g.Q.call(this,{D:"div",
-I:"ytp-remote",K:[{D:"div",I:"ytp-remote-display-status",K:[{D:"div",I:"ytp-remote-display-status-icon",K:[g.QL()]},{D:"div",I:"ytp-remote-display-status-text",W:"{{statustext}}"}]}]});this.u=new g.OL(this,250);g.C(this,this.u);this.w=a;this.L(a,"presentingplayerstatechange",this.A);Wxa(this,g.MK(a))},Wxa=function(a,b){if(3==a.w.getPresentingPlayerType()){var c={RECEIVER_NAME:a.w.getOption("remote","currentReceiver").name};
-c=g.V(b,128)?g.tM("Error on $RECEIVER_NAME",c):b.Ya()||g.V(b,4)?g.tM("Playing on $RECEIVER_NAME",c):g.tM("Connected to $RECEIVER_NAME",c);a.ka("statustext",c);a.u.show()}else a.u.hide()},O8=function(a,b){g.oO.call(this,"Play on",0,a,b);
-this.u=a;this.C={};this.L(a,"onMdxReceiversChange",this.J);this.L(a,"presentingplayerstatechange",this.J);this.J()},P8=function(a){g.sL.call(this,a);
-this.u={key:i6(),name:"This computer"};this.B=null;this.A=[];this.R=this.o=null;this.J=[this.u];this.w=this.u;this.F=new g.MC(64);this.V=0;this.H=-1;this.C=null;if(!g.Xy(this.player.N())){a=this.player;var b=g.gC(a);b&&(b=b.Pl())&&(b=new O8(a,b),g.C(this,b));b=new N8(a);g.C(this,b);g.fL(a,b.element,4);this.C=new M8;g.C(this,this.C);g.fL(a,this.C.element,4);this.M=!!u8()}this.G=null},Q8=function(a){a.G&&(a.player.removeEventListener("presentingplayerstatechange",a.G),a.G=null)},Vxa=function(a,b,c){a.F=
-c;
-a.player.P("presentingplayerstatechange",new g.XC(c,b))},R8=function(a,b){if(b.key!=a.w.key)if(b.key==a.u.key)z8();
-else{var c;(c=!g.R(a.player.N().experiments,"mdx_enable_privacy_disclosure_ui"))||(c=((c=g.L("PLAYER_CONFIG"))&&c.args&&void 0!==c.args.authuser?!0:!(!g.L("SESSION_INDEX")&&!g.L("LOGGED_IN")))||a.M||!a.C);if(c?0:g.fz(a.player.N())||"WEB_EMBEDDED_PLAYER"==a.player.N().deviceParams.c)g.MK(a.player).Ya()?a.player.pauseVideo():(a.G=(0,g.x)(a.RN,a),a.player.addEventListener("presentingplayerstatechange",a.G)),a.C&&a.C.Yb(),x8()||(E8=!0);a.w=b;var d=a.player.getPlaylistId();c=a.player.getVideoData(1);var e=
-c.videoId;if(!d&&!e||(2==a.player.getAppState()||1==a.player.getAppState())&&g.R(a.player.N().experiments,"should_clear_video_data_on_player_cued_unstarted"))c=null;else{var f=a.player.getPlaylist();if(f){var k=[];for(var l=0;l<f.getLength();l++)k[l]=f.va(l).videoId}else k=[e];f=a.player.getCurrentTime(1);d={videoIds:k,listId:d,videoId:e,playerParams:c.playerParams,clickTrackingParams:c.uf,index:Math.max(a.player.getPlaylistIndex(),0),currentTime:0==f?void 0:f};(c=rva(c))&&(d.locationInfo=c);c=d}r8("Connecting to: "+
-g.Lk(b));"cast-selector-receiver"==b.key?(t8(c||null),c=c||null,O7()?L7().setLaunchParams(c):J7("setLaunchParams called before ready.")):!c&&C8()&&q8()==b.key?c6("yt-remote-connection-change",!0):(z8(),t8(c||null),c=p8().le(),(c=k6(c,b.key))&&A8(c,1))}},cwa={"\x00":"\\0",
-"\b":"\\b","\f":"\\f","\n":"\\n","\r":"\\r","\t":"\\t","\x0B":"\\x0B",'"':'\\"',"\\":"\\\\","<":"\\u003C"},I6={"'":"\\'"},Xxa={},Ava={ZP:"atp",FS:"ska",uS:"que",OR:"mus",ES:"sus",aR:"dsp",BS:"seq"};
-d6.prototype.bi=function(){var a=new d6({id:this.id,name:this.name,app:this.app,type:this.type,user:this.username,userAvatarUri:this.avatar,obfuscatedGaiaId:this.obfuscatedGaiaId,theme:this.theme,capabilities:Array.from(this.capabilities.values()).join(","),experiments:Array.from(this.experiments.values()).join(",")});a.w=this.w;a.u=this.u;a.o=this.o};
-var r6,n6="",Sva=Ova("loadCastFramework")||Ova("loadCastApplicationFramework"),mxa=["pkedcjkdefgpdelpbcmbmeomcjbeemfm","enhhojjnijigcajfphajepfemndkmdlo"];/*
-
- Copyright The Closure Library Authors.
- SPDX-License-Identifier: Apache-2.0
-*/
-g.Ua(u6,g.A);g.h=u6.prototype;g.h.nF=function(a){this.A=arguments;this.o=!1;this.Ea?this.w=(0,g.G)()+this.Cd:this.Ea=g.Vf(this.B,this.Cd)};
-g.h.stop=function(){this.Ea&&(g.v.clearTimeout(this.Ea),this.Ea=null);this.w=null;this.o=!1;this.A=[]};
-g.h.pause=function(){++this.u};
-g.h.resume=function(){this.u&&(--this.u,!this.u&&this.o&&(this.o=!1,this.C.apply(null,this.A)))};
-g.h.X=function(){this.stop();u6.Eb.X.call(this)};
-g.h.oF=function(){this.w?(this.Ea=g.Vf(this.B,this.w-(0,g.G)()),this.w=null):(this.Ea=null,this.u?this.o=!0:(this.o=!1,this.C.apply(null,this.A)))};g.Ua(v6,g.A);g.h=v6.prototype;g.h.qj=!1;g.h.fl=0;g.h.Ea=null;g.h.Cx=function(a){this.o=arguments;this.Ea||this.fl?this.qj=!0:w6(this)};
-g.h.stop=function(){this.Ea&&(g.v.clearTimeout(this.Ea),this.Ea=null,this.qj=!1,this.o=[])};
-g.h.pause=function(){this.fl++};
-g.h.resume=function(){this.fl--;this.fl||!this.qj||this.Ea||(this.qj=!1,w6(this))};
-g.h.X=function(){v6.Eb.X.call(this);this.stop()};
-g.h.pF=function(){this.Ea=null;this.qj&&!this.fl&&(this.qj=!1,w6(this))};y6.prototype.stringify=function(a){return g.v.JSON.stringify(a,void 0)};
-y6.prototype.parse=function(a){return g.v.JSON.parse(a,void 0)};z6.prototype.o=null;z6.prototype.getOptions=function(){var a;(a=this.o)||(a={},Wva(this)&&(a[0]=!0,a[1]=!0),a=this.o=a);return a};var S8;g.Ua(Vva,z6);S8=new Vva;g.h=A6.prototype;g.h.ni=null;g.h.bf=!1;g.h.rk=null;g.h.Zu=null;g.h.ak=null;g.h.rj=null;g.h.Pg=null;g.h.Eh=null;g.h.Li=null;g.h.Lc=null;g.h.Lm=0;g.h.gf=null;g.h.fq=null;g.h.fh=null;g.h.jl=-1;g.h.ZB=!0;g.h.Wh=!1;g.h.Vr=0;g.h.zp=null;var awa={},D6={};g.h=A6.prototype;g.h.setTimeout=function(a){this.B=a};
-g.h.cO=function(a){a=a.target;var b=this.zp;b&&3==R6(a)?b.Cx():this.WC(a)};
-g.h.WC=function(a){try{if(a==this.Lc)a:{var b=R6(this.Lc),c=this.Lc.C,d=this.Lc.getStatus();if(g.je&&!g.Nd(10)||g.le&&!g.Md("420+")){if(4>b)break a}else if(3>b||3==b&&!g.yh&&!T6(this.Lc))break a;this.Wh||4!=b||7==c||(8==c||0>=d?this.o.af(3):this.o.af(2));K6(this);var e=this.Lc.getStatus();this.jl=e;var f=T6(this.Lc);(this.bf=200==e)?(4==b&&G6(this),this.C?(bwa(this,b,f),g.yh&&this.bf&&3==b&&(this.A.ia(this.u,"tick",this.VN),this.u.start())):F6(this,f),this.bf&&!this.Wh&&(4==b?this.o.eo(this):(this.bf=
-!1,C6(this)))):(400==e&&0<f.indexOf("Unknown SID")?(this.fh=3,E6(13)):(this.fh=0,E6(14)),G6(this),H6(this))}}catch(k){this.Lc&&T6(this.Lc)}finally{}};
-g.h.VN=function(){var a=R6(this.Lc),b=T6(this.Lc);this.Lm<b.length&&(K6(this),bwa(this,a,b),this.bf&&4!=a&&C6(this))};
-g.h.xN=function(a){J6((0,g.x)(this.wN,this,a),0)};
-g.h.wN=function(a){this.Wh||(K6(this),F6(this,a),C6(this))};
-g.h.hB=function(a){J6((0,g.x)(this.vN,this,a),0)};
-g.h.vN=function(a){this.Wh||(G6(this),this.bf=a,this.o.eo(this),this.o.af(4))};
-g.h.cancel=function(){this.Wh=!0;G6(this)};
-g.h.FN=function(){this.rk=null;var a=(0,g.G)();0<=a-this.Zu?(2!=this.rj&&this.o.af(3),G6(this),this.fh=2,E6(18),H6(this)):ewa(this,this.Zu-a)};
-g.h.getLastError=function(){return this.fh};g.h=hwa.prototype;g.h.Tr=null;g.h.Xd=null;g.h.Ap=!1;g.h.Ex=null;g.h.Nn=null;g.h.Gs=null;g.h.Ur=null;g.h.Ae=null;g.h.lg=-1;g.h.il=null;g.h.Ck=null;g.h.connect=function(a){this.Ur=a;a=M6(this.o,null,this.Ur);E6(3);this.Ex=(0,g.G)();var b=this.o.G;null!=b?(this.il=b[0],(this.Ck=b[1])?(this.Ae=1,iwa(this)):(this.Ae=2,O6(this))):($5(a,"MODE","init"),this.Xd=new A6(this,void 0,void 0,void 0),this.Xd.ni=this.Tr,B6(this.Xd,a,!1,null,!0),this.Ae=0)};
-g.h.jE=function(a){if(a)this.Ae=2,O6(this);else{E6(4);var b=this.o;b.sf=b.Tg.lg;$6(b,9)}a&&this.af(2)};
-g.h.kr=function(a){return this.o.kr(a)};
-g.h.abort=function(){this.Xd&&(this.Xd.cancel(),this.Xd=null);this.lg=-1};
-g.h.Dx=function(){return!1};
-g.h.ZA=function(a,b){this.lg=a.jl;if(0==this.Ae)if(b){try{var c=this.u.parse(b)}catch(d){c=this.o;c.sf=this.lg;$6(c,2);return}this.il=c[0];this.Ck=c[1]}else c=this.o,c.sf=this.lg,$6(c,2);else if(2==this.Ae)if(this.Ap)E6(7),this.Gs=(0,g.G)();else if("11111"==b){if(E6(6),this.Ap=!0,this.Nn=(0,g.G)(),c=this.Nn-this.Ex,!g.je||g.Nd(10)||500>c)this.lg=200,this.Xd.cancel(),E6(12),N6(this.o,this,!0)}else E6(8),this.Nn=this.Gs=(0,g.G)(),this.Ap=!1};
-g.h.eo=function(){this.lg=this.Xd.jl;if(this.Xd.bf)0==this.Ae?this.Ck?(this.Ae=1,iwa(this)):(this.Ae=2,O6(this)):2==this.Ae&&((!g.je||g.Nd(10)?!this.Ap:200>this.Gs-this.Nn)?(E6(11),N6(this.o,this,!1)):(E6(12),N6(this.o,this,!0)));else{0==this.Ae?E6(9):2==this.Ae&&E6(10);var a=this.o;this.Xd.getLastError();a.sf=this.lg;$6(a,2)}};
-g.h.Bm=function(){return this.o.Bm()};
-g.h.isActive=function(){return this.o.isActive()};
-g.h.af=function(a){this.o.af(a)};g.Ua(P6,g.jf);var nwa=/^https?$/i,Yxa=["POST","PUT"];g.h=P6.prototype;
-g.h.send=function(a,b,c,d){if(this.o)throw Error("[goog.net.XhrIo] Object is active with another request="+this.H+"; newUri="+a);b=b?b.toUpperCase():"GET";this.H=a;this.A="";this.C=0;this.da=b;this.T=!1;this.w=!0;this.o=this.R?Xva(this.R):Xva(S8);this.V=this.R?this.R.getOptions():S8.getOptions();this.o.onreadystatechange=(0,g.x)(this.VA,this);try{b6(S6(this,"Opening Xhr")),this.aa=!0,this.o.open(b,String(a),!0),this.aa=!1}catch(f){b6(S6(this,"Error opening Xhr: "+f.message));mwa(this,f);return}a=
-c||"";var e=this.headers.clone();d&&wva(d,function(f,k){e.set(k,f)});
-d=g.Za(e.xe(),kwa);c=g.v.FormData&&a instanceof g.v.FormData;!g.bb(Yxa,b)||d||c||e.set("Content-Type","application/x-www-form-urlencoded;charset=utf-8");e.forEach(function(f,k){this.o.setRequestHeader(k,f)},this);
-this.ea&&(this.o.responseType=this.ea);"withCredentials"in this.o&&this.o.withCredentials!==this.Z&&(this.o.withCredentials=this.Z);try{pwa(this),0<this.F&&(this.M=jwa(this.o),b6(S6(this,"Will abort after "+this.F+"ms if incomplete, xhr2 "+this.M)),this.M?(this.o.timeout=this.F,this.o.ontimeout=(0,g.x)(this.Fx,this)):this.J=g.Vf(this.Fx,this.F,this)),b6(S6(this,"Sending request")),this.G=!0,this.o.send(a),this.G=!1}catch(f){b6(S6(this,"Send error: "+f.message)),mwa(this,f)}};
-g.h.Fx=function(){"undefined"!=typeof g.N1&&this.o&&(this.A="Timed out after "+this.F+"ms, aborting",this.C=8,S6(this,this.A),this.dispatchEvent("timeout"),this.abort(8))};
-g.h.abort=function(a){this.o&&this.w&&(S6(this,"Aborting"),this.w=!1,this.B=!0,this.o.abort(),this.B=!1,this.C=a||7,this.dispatchEvent("complete"),this.dispatchEvent("abort"),Q6(this))};
-g.h.X=function(){this.o&&(this.w&&(this.w=!1,this.B=!0,this.o.abort(),this.B=!1),Q6(this,!0));P6.Eb.X.call(this)};
-g.h.VA=function(){this.ga()||(this.aa||this.G||this.B?owa(this):this.vM())};
-g.h.vM=function(){owa(this)};
-g.h.isActive=function(){return!!this.o};
-g.h.getStatus=function(){try{return 2<R6(this)?this.o.status:-1}catch(a){return-1}};
-g.h.getLastError=function(){return"string"===typeof this.A?this.A:String(this.A)};g.h=U6.prototype;g.h.Rk=null;g.h.Nd=null;g.h.pc=null;g.h.Sr=null;g.h.Qn=null;g.h.dw=null;g.h.co=null;g.h.Xl=0;g.h.RJ=0;g.h.yd=null;g.h.Yg=null;g.h.Vf=null;g.h.di=null;g.h.Tg=null;g.h.cq=null;g.h.Ej=-1;g.h.my=-1;g.h.sf=-1;g.h.kj=0;g.h.cj=0;g.h.Xh=8;var a7=new g.jf;g.Ua(rwa,g.Le);g.Ua(swa,g.Le);g.Ua(twa,g.Le);g.h=U6.prototype;g.h.connect=function(a,b,c,d,e){E6(0);this.Sr=b;this.Rk=c||{};d&&void 0!==e&&(this.Rk.OSID=d,this.Rk.OAID=e);this.C?(J6((0,g.x)(this.zw,this,a),100),xwa(this)):this.zw(a)};
-g.h.zw=function(a){this.Tg=new hwa(this);this.Tg.Tr=null;this.Tg.u=this.B;this.Tg.connect(a)};
-g.h.Dx=function(){return 0==this.o};
-g.h.gB=function(a){this.Yg=null;Awa(this,a)};
-g.h.fB=function(){this.Vf=null;this.pc=new A6(this,this.w,"rpc",this.F);this.pc.ni=null;this.pc.Vr=0;var a=this.dw.clone();Y5(a,"RID","rpc");Y5(a,"SID",this.w);Y5(a,"CI",this.cq?"0":"1");Y5(a,"AID",this.Ej);V6(this,a);if(!g.je||g.Nd(10))Y5(a,"TYPE","xmlhttp"),B6(this.pc,a,!0,this.co,!1);else{Y5(a,"TYPE","html");var b=this.pc,c=!!this.co;b.rj=3;b.Pg=Z5(a.clone());dwa(b,c)}};
-g.h.ZA=function(a,b){if(0!=this.o&&(this.pc==a||this.Nd==a))if(this.sf=a.jl,this.Nd==a&&3==this.o)if(7<this.Xh){try{var c=this.B.parse(b)}catch(f){c=null}if(Array.isArray(c)&&3==c.length)if(0==c[0])a:{if(!this.Vf){if(this.pc)if(this.pc.ak+3E3<this.Nd.ak)X6(this),this.pc.cancel(),this.pc=null;else break a;Z6(this);E6(19)}}else this.my=c[1],0<this.my-this.Ej&&37500>c[2]&&this.cq&&0==this.cj&&!this.di&&(this.di=J6((0,g.x)(this.uK,this),6E3));else $6(this,11)}else b!=Xxa.wQ.o&&$6(this,11);else if(this.pc==
-a&&X6(this),!g.fc(b)){c=this.B.parse(b);for(var d=0;d<c.length;d++){var e=c[d];this.Ej=e[0];e=e[1];2==this.o?"c"==e[0]?(this.w=e[1],this.co=e[2],e=e[3],null!=e?this.Xh=e:this.Xh=6,this.o=3,this.yd&&this.yd.tw(),this.dw=M6(this,this.Bm()?this.co:null,this.Sr),Bwa(this)):"stop"==e[0]&&$6(this,7):3==this.o&&("stop"==e[0]?$6(this,7):"noop"!=e[0]&&this.yd&&this.yd.rw(e),this.cj=0)}}};
-g.h.uK=function(){null!=this.di&&(this.di=null,this.pc.cancel(),this.pc=null,Z6(this),E6(20))};
-g.h.eo=function(a){if(this.pc==a){X6(this);this.pc=null;var b=2}else if(this.Nd==a)this.Nd=null,b=1;else return;this.sf=a.jl;if(0!=this.o)if(a.bf)1==b?(b=(0,g.G)()-a.ak,a7.dispatchEvent(new swa(a7,a.Li?a.Li.length:0,b,this.kj)),W6(this),this.A.length=0):Bwa(this);else{var c=a.getLastError(),d;if(!(d=3==c||7==c||0==c&&0<this.sf)){if(d=1==b)this.Nd||this.Yg||1==this.o||2<=this.kj?d=!1:(this.Yg=J6((0,g.x)(this.gB,this,a),Cwa(this,this.kj)),this.kj++,d=!0);d=!(d||2==b&&Z6(this))}if(d)switch(c){case 1:$6(this,
-5);break;case 4:$6(this,10);break;case 3:$6(this,6);break;case 7:$6(this,12);break;default:$6(this,2)}}};
-g.h.IE=function(a){if(!g.bb(arguments,this.o))throw Error("Unexpected channel state: "+this.o);};
-g.h.OO=function(a){a?E6(2):(E6(1),Dwa(this,8))};
-g.h.kr=function(a){if(a)throw Error("Can't create secondary domain capable XhrIo object.");a=new P6;a.Z=!1;return a};
-g.h.isActive=function(){return!!this.yd&&this.yd.isActive(this)};
-g.h.af=function(a){a7.dispatchEvent(new twa(a7,a))};
-g.h.Bm=function(){return!(!g.je||g.Nd(10))};
-g.h=Ewa.prototype;g.h.tw=function(){};
-g.h.rw=function(){};
-g.h.qw=function(){};
-g.h.Zq=function(){};
-g.h.ax=function(){return{}};
-g.h.isActive=function(){return!0};g.h=Fwa.prototype;g.h.isEmpty=function(){return g.cb(this.o)&&g.cb(this.u)};
-g.h.clear=function(){this.o=[];this.u=[]};
-g.h.contains=function(a){return g.bb(this.o,a)||g.bb(this.u,a)};
-g.h.remove=function(a){var b=this.o;var c=(0,g.vqa)(b,a);0<=c?(g.eb(b,c),b=!0):b=!1;return b||g.fb(this.u,a)};
-g.h.Gd=function(){for(var a=[],b=this.o.length-1;0<=b;--b)a.push(this.o[b]);var c=this.u.length;for(b=0;b<c;++b)a.push(this.u[b]);return a};g.Ua(b7,g.A);g.h=b7.prototype;g.h.oN=function(){this.Cd=Math.min(3E5,2*this.Cd);this.w();this.u&&this.start()};
-g.h.start=function(){var a=this.Cd+15E3*Math.random();g.An(this.o,a);this.u=(0,g.G)()+a};
-g.h.stop=function(){this.o.stop();this.u=0};
-g.h.isActive=function(){return this.o.isActive()};
-g.h.reset=function(){this.o.stop();this.Cd=5E3};g.Ua(c7,Ewa);g.h=c7.prototype;g.h.subscribe=function(a,b,c){return this.w.subscribe(a,b,c)};
-g.h.unsubscribe=function(a,b,c){return this.w.unsubscribe(a,b,c)};
-g.h.uh=function(a){return this.w.oi(a)};
-g.h.P=function(a,b){return this.w.P.apply(this.w,arguments)};
-g.h.dispose=function(){this.vb||(this.vb=!0,g.Je(this.w),Iwa(this),g.Je(this.u),this.u=null)};
-g.h.ga=function(){return this.vb};
-g.h.connect=function(a,b,c){if(!this.o||2!=this.o.o){this.H="";this.u.stop();this.B=a||null;this.A=b||0;a=this.J+"/test";b=this.J+"/bind";var d=new U6(c?c.firstTestResults:null,c?c.secondTestResults:null,this.M),e=this.o;e&&(e.yd=null);d.yd=this;this.o=d;e?this.o.connect(a,b,this.C,e.w,e.Ej):c?this.o.connect(a,b,this.C,c.sessionId,c.arrayId):this.o.connect(a,b,this.C)}};
-g.h.sendMessage=function(a,b){var c={_sc:a};b&&g.Sb(c,b);this.u.isActive()||2==(this.o?this.o.o:0)?this.F.push(c):d7(this)&&ywa(this.o,c)};
-g.h.tw=function(){this.u.reset();this.B=null;this.A=0;if(this.F.length){var a=this.F;this.F=[];for(var b=0,c=a.length;b<c;++b)ywa(this.o,a[b])}this.P("handlerOpened")};
-g.h.qw=function(a){var b=2==a&&401==this.o.sf;4==a||b||this.u.start();this.P("handlerError",a)};
-g.h.Zq=function(a){if(!this.u.isActive())this.P("handlerClosed");else if(a)for(var b=0,c=a.length;b<c;++b){var d=a[b].map;d&&this.F.push(d)}};
-g.h.ax=function(){var a={v:2};this.H&&(a.gsessionid=this.H);0!=this.A&&(a.ui=""+this.A);0!=this.G&&(a.ui=""+this.G);this.B&&g.Sb(a,this.B);return a};
-g.h.rw=function(a){"S"==a[0]?this.H=a[1]:"gracefulReconnect"==a[0]?(this.u.start(),wwa(this.o)):this.P("handlerMessage",new Gwa(a[0],a[1]))};
-g.h.qO=function(){this.u.isActive();var a=this.o,b=0;a.pc&&b++;a.Nd&&b++;0==b&&this.connect(this.B,this.A)};e7.prototype.A=function(a,b,c,d){b?a(d):a({text:c.responseText})};
-e7.prototype.w=function(a,b){a(Error("Request error: "+b.status))};
-e7.prototype.B=function(a){a(Error("request timed out"))};var Qwa=(0,g.G)(),h7=null,k7=Array(50),j7=-1,l7=!1;g.Ua(m7,g.P);m7.prototype.le=function(){return this.o};
-m7.prototype.contains=function(a){return!!j6(this.o,a)};
-m7.prototype.get=function(a){return a?k6(this.o,a):null};
-m7.prototype.info=function(a){i7(this.F,a)};g.r(n7,g.P);g.h=n7.prototype;g.h.start=function(){!this.o&&isNaN(this.Ea)&&this.aD()};
-g.h.stop=function(){this.o&&(this.o.abort(),this.o=null);isNaN(this.Ea)||(g.No(this.Ea),this.Ea=NaN)};
-g.h.X=function(){this.stop();g.P.prototype.X.call(this)};
-g.h.aD=function(){this.Ea=NaN;this.o=g.Iq(f7(this.C,"/pairing/get_screen"),{method:"POST",ub:{pairing_code:this.A},timeout:5E3,onSuccess:(0,g.x)(this.uP,this),onError:(0,g.x)(this.tP,this),Ud:(0,g.x)(this.vP,this)})};
-g.h.uP=function(a,b){this.o=null;var c=b.screen||{};c.dialId=this.B;c.name=this.w;this.P("pairingComplete",new e6(c))};
-g.h.tP=function(a){this.o=null;a.status&&404==a.status?this.u>=Zxa.length?this.P("pairingFailed",Error("DIAL polling timed out")):(a=Zxa[this.u],this.Ea=g.Lo((0,g.x)(this.aD,this),a),this.u++):this.P("pairingFailed",Error("Server error "+a.status))};
-g.h.vP=function(){this.o=null;this.P("pairingFailed",Error("Server not responding"))};
-var Zxa=[2E3,2E3,1E3,1E3,1E3,2E3,2E3,5E3,5E3,1E4];g.Ua(p7,m7);g.h=p7.prototype;g.h.start=function(){o7(this)&&this.P("screenChange");!g.Hs("yt-remote-lounge-token-expiration")&&Uwa(this);g.No(this.u);this.u=g.Lo((0,g.x)(this.start,this),1E4)};
-g.h.add=function(a,b){o7(this);Rwa(this,a);q7(this,!1);this.P("screenChange");b(a);a.token||Uwa(this)};
-g.h.remove=function(a,b){var c=o7(this);Twa(this,a)&&(q7(this,!1),c=!0);b(a);c&&this.P("screenChange")};
-g.h.Zp=function(a,b,c,d){var e=o7(this),f=this.get(a.id);f?(f.name!=b&&(f.name=b,q7(this,!1),e=!0),c(a)):d(Error("no such local screen."));e&&this.P("screenChange")};
-g.h.X=function(){g.No(this.u);p7.Eb.X.call(this)};
-g.h.zF=function(a){o7(this);var b=this.o.length;a=a&&a.screens||[];for(var c=0,d=a.length;c<d;++c){var e=a[c],f=this.get(e.screenId);f&&(f.token=e.loungeToken,--b)}q7(this,!b);b&&i7(this.F,"Missed "+b+" lounge tokens.")};
-g.h.yF=function(a){i7(this.F,"Requesting lounge tokens failed: "+a)};g.r(s7,g.P);g.h=s7.prototype;g.h.start=function(){var a=parseInt(g.Hs("yt-remote-fast-check-period")||"0",10);(this.A=(0,g.G)()-144E5<a?0:a)?u7(this):(this.A=(0,g.G)()+3E5,g.Gs("yt-remote-fast-check-period",this.A),this.qu())};
-g.h.isEmpty=function(){return g.Mb(this.o)};
-g.h.update=function(){r7("Updating availability on schedule.");var a=this.C(),b=g.Ab(this.o,function(c,d){return c&&!!k6(a,d)},this);
-t7(this,b)};
-g.h.X=function(){g.No(this.w);this.w=NaN;this.u&&(this.u.abort(),this.u=null);g.P.prototype.X.call(this)};
-g.h.qu=function(){g.No(this.w);this.w=NaN;this.u&&this.u.abort();var a=Xwa(this);if(a6(a)){var b=f7(this.B,"/pairing/get_screen_availability");this.u=g7(this.B,b,{lounge_token:g.Gb(a).join(",")},(0,g.x)(this.SM,this,a),(0,g.x)(this.RM,this))}else t7(this,{}),u7(this)};
-g.h.SM=function(a,b){this.u=null;var c=g.Gb(Xwa(this));if(g.tb(c,g.Gb(a))){c=b.screens||[];for(var d={},e=0,f=c.length;e<f;++e)d[a[c[e].loungeToken]]="online"==c[e].status;t7(this,d);u7(this)}else this.Sb("Changing Screen set during request."),this.qu()};
-g.h.RM=function(a){this.Sb("Screen availability failed: "+a);this.u=null;u7(this)};
-g.h.Sb=function(a){i7("OnlineScreenService",a)};g.Ua(v7,m7);g.h=v7.prototype;g.h.start=function(){this.w.start();this.u.start();this.o.length&&(this.P("screenChange"),this.u.isEmpty()||this.P("onlineScreenChange"))};
-g.h.add=function(a,b,c){this.w.add(a,b,c)};
-g.h.remove=function(a,b,c){this.w.remove(a,b,c);this.u.update()};
-g.h.Zp=function(a,b,c,d){this.w.contains(a)?this.w.Zp(a,b,c,d):(a="Updating name of unknown screen: "+a.name,i7(this.F,a),d(Error(a)))};
-g.h.le=function(a){return a?this.o:g.ib(this.o,(0,g.we)(this.A,function(b){return!this.contains(b)},this))};
-g.h.bD=function(){return(0,g.we)(this.le(!0),function(a){return!!this.u.o[a.id]},this)};
-g.h.cD=function(a,b,c,d,e){this.info("getDialScreenByPairingCode "+a+" / "+b);var f=new n7(this.C,a,b,c);f.subscribe("pairingComplete",(0,g.x)(function(k){g.Je(f);d(w7(this,k))},this));
-f.subscribe("pairingFailed",function(k){g.Je(f);e(k)});
-f.start();return(0,g.x)(f.stop,f)};
-g.h.xP=function(a,b,c,d){g.Iq(f7(this.C,"/pairing/get_screen"),{method:"POST",ub:{pairing_code:a},timeout:5E3,onSuccess:(0,g.x)(function(e,f){var k=new e6(f.screen||{});if(!k.name||axa(this,k.name)){a:{var l=k.name;for(var m=2,n=b(l,m);axa(this,n);){m++;if(20<m)break a;n=b(l,m)}l=n}k.name=l}c(w7(this,k))},this),
-onError:(0,g.x)(function(e){d(Error("pairing request failed: "+e.status))},this),
-Ud:(0,g.x)(function(){d(Error("pairing request timed out."))},this)})};
-g.h.X=function(){g.Je(this.w);g.Je(this.u);v7.Eb.X.call(this)};
-g.h.IF=function(){bxa(this);this.P("screenChange");this.u.update()};
-v7.prototype.dispose=v7.prototype.dispose;g.Ua(y7,g.P);g.h=y7.prototype;g.h.np=function(a){this.w=a;this.P("sessionScreen",this.w)};
-g.h.ne=function(a){this.ga()||(a&&z7(this,""+a),this.w=null,this.P("sessionScreen",null))};
-g.h.info=function(a){i7(this.V,a)};
-g.h.eD=function(){return null};
-g.h.Au=function(a){var b=this.u;a?(b.displayStatus=new chrome.cast.ReceiverDisplayStatus(a,[]),b.displayStatus.showStop=!0):b.displayStatus=null;chrome.cast.setReceiverDisplayStatus(b,(0,g.x)(function(){this.info("Updated receiver status for "+b.friendlyName+": "+a)},this),(0,g.x)(function(){z7(this,"Failed to update receiver status for: "+b.friendlyName)},this))};
-g.h.X=function(){this.Au("");y7.Eb.X.call(this)};g.Ua(A7,y7);g.h=A7.prototype;g.h.zu=function(a){if(this.o){if(this.o==a)return;z7(this,"Overriding cast sesison with new session object");this.o.removeUpdateListener(this.C);this.o.removeMessageListener("urn:x-cast:com.google.youtube.mdx",this.B)}this.o=a;this.o.addUpdateListener(this.C);this.o.addMessageListener("urn:x-cast:com.google.youtube.mdx",this.B);dxa(this)};
-g.h.Fj=function(a){this.info("launchWithParams no-op for Cast: "+g.Lk(a))};
-g.h.stop=function(){this.o?this.o.stop((0,g.x)(function(){this.ne()},this),(0,g.x)(function(){this.ne(Error("Failed to stop receiver app."))},this)):this.ne(Error("Stopping cast device witout session."))};
-g.h.Au=g.Ja;g.h.X=function(){this.info("disposeInternal");g.No(this.A);this.A=0;this.o&&(this.o.removeUpdateListener(this.C),this.o.removeMessageListener("urn:x-cast:com.google.youtube.mdx",this.B));this.o=null;A7.Eb.X.call(this)};
-g.h.NN=function(a,b){if(!this.ga())if(b){var c=x6(b);if(g.Pa(c)){var d=""+c.type;c=c.data||{};this.info("onYoutubeMessage_: "+d+" "+g.Lk(c));switch(d){case "mdxSessionStatus":cxa(this,c.screenId);break;default:z7(this,"Unknown youtube message: "+d)}}else z7(this,"Unable to parse message.")}else z7(this,"No data in message.")};
-g.h.wx=function(a,b,c,d){$wa(this.H,this.u.label,a,this.u.friendlyName,(0,g.x)(function(e){e?b(e):0<=d?(z7(this,"Screen "+a+" appears to be offline. "+d+" retries left."),g.Lo((0,g.x)(this.wx,this,a,b,c,d-1),300)):c(Error("Unable to fetch screen."))},this),c)};
-g.h.eD=function(){return this.o};
-g.h.yP=function(a){this.ga()||a||(z7(this,"Cast session died."),this.ne())};g.Ua(B7,y7);g.h=B7.prototype;g.h.zu=function(a){this.A=a;this.A.addUpdateListener(this.M)};
-g.h.Fj=function(a){this.B=a;this.F()};
-g.h.stop=function(){this.o();this.o=g.Ja;g.No(this.C);this.A?this.A.stop((0,g.x)(this.ne,this,null),(0,g.x)(this.ne,this,"Failed to stop DIAL device.")):this.ne()};
-g.h.X=function(){this.o();this.o=g.Ja;g.No(this.C);this.A&&this.A.removeUpdateListener(this.M);this.A=null;B7.Eb.X.call(this)};
-g.h.BP=function(a){this.ga()||a||(z7(this,"DIAL session died."),this.o(),this.o=g.Ja,this.ne())};
-g.h.ks=function(a){this.J=i6();if(this.B){var b=new chrome.cast.DialLaunchResponse(!0,fxa(this));a(b);exa(this)}else this.F=(0,g.x)(function(){g.No(this.C);this.F=g.Ja;this.C=NaN;var c=new chrome.cast.DialLaunchResponse(!0,fxa(this));a(c);exa(this)},this),this.C=g.Lo((0,g.x)(function(){this.F()},this),100)};
-g.h.bG=function(a,b,c){$wa(this.H,this.G.receiver.label,a,this.u.friendlyName,(0,g.x)(function(d){d&&d.token?(this.np(d),b(new chrome.cast.DialLaunchResponse(!1))):this.ks(b,c)},this),(0,g.x)(function(d){z7(this,"Failed to get DIAL screen: "+d);
-this.ks(b,c)},this))};g.Ua(C7,y7);C7.prototype.stop=function(){this.ne()};
-C7.prototype.zu=g.Ja;C7.prototype.Fj=function(){g.No(this.o);this.o=NaN;var a=k6(this.H.le(),this.u.label);a?this.np(a):this.ne(Error("No such screen"))};
-C7.prototype.X=function(){g.No(this.o);this.o=NaN;C7.Eb.X.call(this)};g.Ua(D7,g.P);g.h=D7.prototype;
-g.h.init=function(a,b){chrome.cast.timeout.requestSession=3E4;var c=new chrome.cast.SessionRequest(this.F);this.G||(c.dialRequest=new chrome.cast.DialRequest("YouTube"));var d=chrome.cast.AutoJoinPolicy.TAB_AND_ORIGIN_SCOPED,e=a?chrome.cast.DefaultActionPolicy.CAST_THIS_TAB:chrome.cast.DefaultActionPolicy.CREATE_SESSION,f=(0,g.x)(this.AM,this);c=new chrome.cast.ApiConfig(c,(0,g.x)(this.cB,this),f,d,e);c.customDialLaunchCallback=(0,g.x)(this.SK,this);chrome.cast.initialize(c,(0,g.x)(function(){this.ga()||
-(chrome.cast.addReceiverActionListener(this.A),Nwa(),this.u.subscribe("onlineScreenChange",(0,g.x)(this.dD,this)),this.w=hxa(this),chrome.cast.setCustomReceivers(this.w,g.Ja,(0,g.x)(function(k){this.Sb("Failed to set initial custom receivers: "+g.Lk(k))},this)),this.P("yt-remote-cast2-availability-change",F7(this)),b(!0))},this),(0,g.x)(function(k){this.Sb("Failed to initialize API: "+g.Lk(k));
-b(!1)},this))};
-g.h.yO=function(a,b){E7("Setting connected screen ID: "+a+" -> "+b);if(this.o){var c=this.o.w;if(!a||c&&c.id!=a)E7("Unsetting old screen status: "+this.o.u.friendlyName),G7(this,null)}if(a&&b){if(!this.o){c=k6(this.u.le(),a);if(!c){E7("setConnectedScreenStatus: Unknown screen.");return}var d=gxa(this,c);d||(E7("setConnectedScreenStatus: Connected receiver not custom..."),d=new chrome.cast.Receiver(c.uuid?c.uuid:c.id,c.name),d.receiverType=chrome.cast.ReceiverType.CUSTOM,this.w.push(d),chrome.cast.setCustomReceivers(this.w,
-g.Ja,(0,g.x)(function(e){this.Sb("Failed to set initial custom receivers: "+g.Lk(e))},this)));
-E7("setConnectedScreenStatus: new active receiver: "+d.friendlyName);G7(this,new C7(this.u,d),!0)}this.o.Au(b)}else E7("setConnectedScreenStatus: no screen.")};
-g.h.zO=function(a){this.ga()?this.Sb("Setting connection data on disposed cast v2"):this.o?this.o.Fj(a):this.Sb("Setting connection data without a session")};
-g.h.AP=function(){this.ga()?this.Sb("Stopping session on disposed cast v2"):this.o?(this.o.stop(),G7(this,null)):E7("Stopping non-existing session")};
-g.h.requestSession=function(){chrome.cast.requestSession((0,g.x)(this.cB,this),(0,g.x)(this.WM,this))};
-g.h.X=function(){this.u.unsubscribe("onlineScreenChange",(0,g.x)(this.dD,this));window.chrome&&chrome.cast&&chrome.cast.removeReceiverActionListener(this.A);var a=Kwa,b=g.w("yt.mdx.remote.debug.handlers_");g.fb(b||[],a);g.Je(this.o);D7.Eb.X.call(this)};
-g.h.Sb=function(a){i7("Controller",a)};
-g.h.eB=function(a,b){this.o==a&&(b||G7(this,null),this.P("yt-remote-cast2-session-change",b))};
-g.h.wM=function(a,b){if(!this.ga())if(a)switch(a.friendlyName=chrome.cast.unescape(a.friendlyName),E7("onReceiverAction_ "+a.label+" / "+a.friendlyName+"-- "+b),b){case chrome.cast.ReceiverAction.CAST:if(this.o)if(this.o.u.label!=a.label)E7("onReceiverAction_: Stopping active receiver: "+this.o.u.friendlyName),this.o.stop();else{E7("onReceiverAction_: Casting to active receiver.");this.o.w&&this.P("yt-remote-cast2-session-change",this.o.w);break}switch(a.receiverType){case chrome.cast.ReceiverType.CUSTOM:G7(this,
-new C7(this.u,a));break;case chrome.cast.ReceiverType.DIAL:G7(this,new B7(this.u,a,this.C));break;case chrome.cast.ReceiverType.CAST:G7(this,new A7(this.u,a));break;default:this.Sb("Unknown receiver type: "+a.receiverType)}break;case chrome.cast.ReceiverAction.STOP:this.o&&this.o.u.label==a.label?this.o.stop():this.Sb("Stopping receiver w/o session: "+a.friendlyName)}else this.Sb("onReceiverAction_ called without receiver.")};
-g.h.SK=function(a){if(this.ga())return Promise.reject(Error("disposed"));var b=a.receiver;b.receiverType!=chrome.cast.ReceiverType.DIAL&&(this.Sb("Not DIAL receiver: "+b.friendlyName),b.receiverType=chrome.cast.ReceiverType.DIAL);var c=this.o?this.o.u:null;if(!c||c.label!=b.label)return this.Sb("Receiving DIAL launch request for non-clicked DIAL receiver: "+b.friendlyName),Promise.reject(Error("illegal DIAL launch"));if(c&&c.label==b.label&&c.receiverType!=chrome.cast.ReceiverType.DIAL){if(this.o.w)return E7("Reselecting dial screen."),
-this.P("yt-remote-cast2-session-change",this.o.w),Promise.resolve(new chrome.cast.DialLaunchResponse(!1));this.Sb('Changing CAST intent from "'+c.receiverType+'" to "dial" for '+b.friendlyName);G7(this,new B7(this.u,b,this.C))}b=this.o;b.G=a;return b.G.appState==chrome.cast.DialAppState.RUNNING?new Promise((0,g.x)(b.bG,b,(b.G.extraData||{}).screenId||null)):new Promise((0,g.x)(b.ks,b))};
-g.h.cB=function(a){if(!this.ga()){E7("New cast session ID: "+a.sessionId);var b=a.receiver;if(b.receiverType!=chrome.cast.ReceiverType.CUSTOM){if(!this.o)if(b.receiverType==chrome.cast.ReceiverType.CAST)E7("Got resumed cast session before resumed mdx connection."),b.friendlyName=chrome.cast.unescape(b.friendlyName),G7(this,new A7(this.u,b),!0);else{this.Sb("Got non-cast session without previous mdx receiver event, or mdx resume.");return}var c=this.o.u,d=k6(this.u.le(),c.label);d&&f6(d,b.label)&&
-c.receiverType!=chrome.cast.ReceiverType.CAST&&b.receiverType==chrome.cast.ReceiverType.CAST&&(E7("onSessionEstablished_: manual to cast session change "+b.friendlyName),g.Je(this.o),this.o=new A7(this.u,b),this.o.subscribe("sessionScreen",(0,g.x)(this.eB,this,this.o)),this.o.Fj(null));this.o.zu(a)}}};
-g.h.zP=function(){return this.o?this.o.eD():null};
-g.h.WM=function(a){this.ga()||(this.Sb("Failed to estabilish a session: "+g.Lk(a)),a.code!=chrome.cast.ErrorCode.CANCEL&&G7(this,null))};
-g.h.AM=function(a){E7("Receiver availability updated: "+a);if(!this.ga()){var b=F7(this);this.B=a==chrome.cast.ReceiverAvailability.AVAILABLE;F7(this)!=b&&this.P("yt-remote-cast2-availability-change",F7(this))}};
-g.h.dD=function(){this.ga()||(this.w=hxa(this),E7("Updating custom receivers: "+g.Lk(this.w)),chrome.cast.setCustomReceivers(this.w,g.Ja,(0,g.x)(function(){this.Sb("Failed to set custom receivers.")},this)),this.P("yt-remote-cast2-availability-change",F7(this)))};
-D7.prototype.setLaunchParams=D7.prototype.zO;D7.prototype.setConnectedScreenStatus=D7.prototype.yO;D7.prototype.stopSession=D7.prototype.AP;D7.prototype.getCastSession=D7.prototype.zP;D7.prototype.requestSession=D7.prototype.requestSession;D7.prototype.init=D7.prototype.init;D7.prototype.dispose=D7.prototype.dispose;var M7=[];g.h=R7.prototype;
-g.h.reset=function(a){this.listId="";this.index=-1;this.videoId="";S7(this);this.volume=-1;this.muted=!1;a&&(this.index=a.index,this.listId=a.listId,this.videoId=a.videoId,this.playerState=a.playerState,this.volume=a.volume,this.muted=a.muted,this.audioTrackId=a.audioTrackId,this.o=a.trackData,this.J=a.hasPrevious,this.hasNext=a.hasNext,this.F=a.playerTime,this.C=a.playerTimeAt,this.B=a.seekableStart,this.u=a.seekableEnd,this.G=a.duration,this.H=a.loadedTime,this.A=a.liveIngestionTime,this.w=!isNaN(this.A))};
-g.h.Ya=function(){return 1==this.playerState};
-g.h.isAdPlaying=function(){return 1081==this.playerState};
-g.h.getDuration=function(){return this.w?this.G+T7(this):this.G};
-g.h.clone=function(){return new R7(X7(this))};g.r(Z7,g.P);g.h=Z7.prototype;g.h.play=function(){1==this.o?(this.u?this.u.play(null,g.Ja,e8(this,"play")):d8(this,"play"),c8(this,1,V7($7(this))),this.P("remotePlayerChange")):a8(this,this.play)};
-g.h.pause=function(){1==this.o?(this.u?this.u.pause(null,g.Ja,e8(this,"pause")):d8(this,"pause"),c8(this,2,V7($7(this))),this.P("remotePlayerChange")):a8(this,this.pause)};
-g.h.seekTo=function(a){if(1==this.o){if(this.u){var b=$7(this),c=new chrome.cast.media.SeekRequest;c.currentTime=a;b.Ya()||3==b.playerState?c.resumeState=chrome.cast.media.ResumeState.PLAYBACK_START:c.resumeState=chrome.cast.media.ResumeState.PLAYBACK_PAUSE;this.u.seek(c,g.Ja,e8(this,"seekTo",{newTime:a}))}else d8(this,"seekTo",{newTime:a});c8(this,3,a);this.P("remotePlayerChange")}else a8(this,g.Sa(this.seekTo,a))};
-g.h.stop=function(){if(1==this.o){this.u?this.u.stop(null,g.Ja,e8(this,"stopVideo")):d8(this,"stopVideo");var a=$7(this);a.index=-1;a.videoId="";S7(a);b8(this,a);this.P("remotePlayerChange")}else a8(this,this.stop)};
-g.h.setVolume=function(a,b){if(1==this.o){var c=$7(this);if(this.w){if(c.volume!=a){var d=Math.round(a)/100;this.w.setReceiverVolumeLevel(d,(0,g.x)(function(){i7("CP","set receiver volume: "+d)},this),(0,g.x)(function(){this.Sb("failed to set receiver volume.")},this))}c.muted!=b&&this.w.setReceiverMuted(b,(0,g.x)(function(){i7("CP","set receiver muted: "+b)},this),(0,g.x)(function(){this.Sb("failed to set receiver muted.")},this))}else{var e={volume:a,
-muted:b};-1!=c.volume&&(e.delta=a-c.volume);d8(this,"setVolume",e)}c.muted=b;c.volume=a;b8(this,c)}else a8(this,g.Sa(this.setVolume,a,b))};
-g.h.eC=function(a,b){if(1==this.o){var c=$7(this),d={videoId:a};b&&(c.o={trackName:b.name,languageCode:b.languageCode,sourceLanguageCode:b.translationLanguage?b.translationLanguage.languageCode:"",languageName:b.languageName,kind:b.kind},d.style=g.Lk(b.style),g.Sb(d,c.o));d8(this,"setSubtitlesTrack",d);b8(this,c)}else a8(this,g.Sa(this.eC,a,b))};
-g.h.setAudioTrack=function(a,b){if(1==this.o){var c=b.getLanguageInfo().getId();d8(this,"setAudioTrack",{videoId:a,audioTrackId:c});var d=$7(this);d.audioTrackId=c;b8(this,d)}else a8(this,g.Sa(this.setAudioTrack,a,b))};
-g.h.playVideo=function(a,b,c,d,e,f,k){var l=$7(this);c=c||0;var m={videoId:a,currentIndex:c};W7(l,a,c);void 0!==b&&(U7(l,b),m.currentTime=b);void 0!==d&&(m.listId=d);null!=e&&(m.playerParams=e);null!=f&&(m.clickTrackingParams=f);null!=k&&(m.locationInfo=g.Lk(k));d8(this,"setPlaylist",m);d||b8(this,l)};
-g.h.ZC=function(a,b){if(1==this.o){if(a&&b){var c=$7(this);W7(c,a,b);b8(this,c)}d8(this,"previous")}else a8(this,g.Sa(this.ZC,a,b))};
-g.h.nextVideo=function(a,b){if(1==this.o){if(a&&b){var c=$7(this);W7(c,a,b);b8(this,c)}d8(this,"next")}else a8(this,g.Sa(this.nextVideo,a,b))};
-g.h.Hw=function(){1==this.o?d8(this,"dismissAutoplay"):a8(this,this.Hw)};
-g.h.dispose=function(){if(3!=this.o){var a=this.o;this.o=3;this.P("proxyStateChange",a,this.o)}g.P.prototype.dispose.call(this)};
-g.h.X=function(){uxa(this);this.A=null;this.B.clear();Y7(this,null);g.P.prototype.X.call(this)};
-g.h.cv=function(a){if((a!=this.o||2==a)&&3!=this.o&&0!=a){var b=this.o;this.o=a;this.P("proxyStateChange",b,a);if(1==a)for(;!this.B.isEmpty();)b=a=this.B,g.cb(b.o)&&(b.o=b.u,b.o.reverse(),b.u=[]),a.o.pop().apply(this);else 3==a&&this.dispose()}};
-g.h.tM=function(a,b){this.P(a,b)};
-g.h.AK=function(a){if(!a)this.Zl(null),Y7(this,null);else if(this.w.receiver.volume){a=this.w.receiver.volume;var b=$7(this),c=Math.round(100*a.level||0);if(b.volume!=c||b.muted!=a.muted)i7("CP","Cast volume update: "+a.level+(a.muted?" muted":"")),b.volume=c,b.muted=!!a.muted,b8(this,b)}};
-g.h.Zl=function(a){i7("CP","Cast media: "+!!a);this.u&&this.u.removeUpdateListener(this.G);if(this.u=a)this.u.addUpdateListener(this.G),vxa(this),this.P("remotePlayerChange")};
-g.h.zK=function(a){a?(vxa(this),this.P("remotePlayerChange")):this.Zl(null)};
-g.h.LK=function(){var a=rxa();a&&Y7(this,a)};
-g.h.Sb=function(a){i7("CP",a)};g.r(f8,g.P);g.h=f8.prototype;
-g.h.connect=function(a,b){if(b){var c=b.listId,d=b.videoId,e=b.playerParams,f=b.clickTrackingParams,k=b.index,l={videoId:d},m=b.currentTime,n=b.locationInfo;void 0!==m&&(l.currentTime=5>=m?0:m);e&&(l.playerParams=e);n&&(l.locationInfo=n);f&&(l.clickTrackingParams=f);c&&(l.listId=c);void 0!==k&&(l.currentIndex=k);c&&(this.Sa.listId=c);this.Sa.videoId=d;this.Sa.index=k||0;this.Sa.state=3;U7(this.Sa,m);this.A="UNSUPPORTED";g8("Connecting with setPlaylist and params: "+g.Lk(l));this.o.connect({method:"setPlaylist",
-params:g.Lk(l)},a,Iva())}else g8("Connecting without params"),this.o.connect({},a,Iva());xxa(this)};
-g.h.dispose=function(){this.ga()||(this.P("beforeDispose"),h8(this,3));g.P.prototype.dispose.call(this)};
-g.h.X=function(){i8(this);k8(this);j8(this);g.No(this.F);this.F=NaN;g.No(this.G);this.G=NaN;this.w=null;g.Jp(this.V);this.V.length=0;this.o.dispose();g.P.prototype.X.call(this);this.A=this.C=this.u=this.Sa=this.o=null};
-g.h.rF=function(){this.ij(2)};
-g.h.EK=function(){g8("Channel opened");this.M&&(this.M=!1,j8(this),this.J=g.Lo((0,g.x)(function(){g8("Timing out waiting for a screen.");this.ij(1)},this),15E3));
-Mva(Hwa(this.o),this.R)};
-g.h.BK=function(){g8("Channel closed");isNaN(this.B)?q6(!0):q6();this.dispose()};
-g.h.CK=function(a){q6();isNaN(this.Yk())?(g8("Channel error: "+a+" without reconnection"),this.dispose()):(this.M=!0,g8("Channel error: "+a+" with reconnection in "+this.Yk()+" ms"),h8(this,2))};
-g.h.DK=function(a){a.params?g8("Received: action="+a.action+", params="+g.Lk(a.params)):g8("Received: action="+a.action+" {}");switch(a.action){case "loungeStatus":a=x6(a.params.devices);this.u=(0,g.Ec)(a,function(c){return new d6(c)});
-a=!!g.Za(this.u,function(c){return"LOUNGE_SCREEN"==c.type});
-Axa(this,a);break;case "loungeScreenDisconnected":g.gb(this.u,function(c){return"LOUNGE_SCREEN"==c.type});
-Axa(this,!1);break;case "remoteConnected":var b=new d6(x6(a.params.device));g.Za(this.u,function(c){return b?c.id==b.id:!1})||sva(this.u,b);
-break;case "remoteDisconnected":b=new d6(x6(a.params.device));g.gb(this.u,function(c){return b?c.id==b.id:!1});
-break;case "gracefulDisconnect":break;case "playlistModified":Cxa(this,a);break;case "nowPlaying":Exa(this,a);break;case "onStateChange":Dxa(this,a);break;case "onAdStateChange":Fxa(this,a);break;case "onVolumeChanged":Gxa(this,a);break;case "onSubtitlesTrackChanged":Bxa(this,a);break;case "nowAutoplaying":Hxa(this,a);break;case "autoplayDismissed":this.P("autoplayDismissed");break;case "autoplayUpNext":this.C=a.params.videoId||null;this.P("autoplayUpNext",this.C);break;case "onAutoplayModeChanged":this.A=
-a.params.autoplayMode;this.P("autoplayModeChange",this.A);"DISABLED"==this.A&&this.P("autoplayDismissed");break;case "onHasPreviousNextChanged":Ixa(this,a);break;case "requestAssistedSignIn":this.P("assistedSignInRequested",a.params.authCode);break;default:g8("Unrecognized action: "+a.action)}};
-g.h.mO=function(){if(this.w){var a=this.w;this.w=null;this.Sa.videoId!=a&&l8(this,"getNowPlaying")}};
-g.h.gF=function(){var a=3;this.ga()||(a=0,isNaN(this.Yk())?d7(this.o)&&isNaN(this.B)&&(a=1):a=2);return a};
-g.h.ij=function(a){g8("Disconnecting with "+a);i8(this);this.P("beforeDisconnect",a);1==a&&q6();Iwa(this.o,a);this.dispose()};
-g.h.fF=function(){var a=this.Sa;this.w&&(a=this.Sa.clone(),W7(a,this.w,a.index));return X7(a)};
-g.h.AO=function(a){var b=new R7(a);b.videoId&&b.videoId!=this.Sa.videoId&&(this.w=b.videoId,g.No(this.F),this.F=g.Lo((0,g.x)(this.mO,this),5E3));var c=[];this.Sa.listId==b.listId&&this.Sa.videoId==b.videoId&&this.Sa.index==b.index||c.push("remoteQueueChange");this.Sa.playerState==b.playerState&&this.Sa.volume==b.volume&&this.Sa.muted==b.muted&&V7(this.Sa)==V7(b)&&g.Lk(this.Sa.o)==g.Lk(b.o)||c.push("remotePlayerChange");this.Sa.reset(a);(0,g.y)(c,function(d){this.P(d)},this)};
-g.h.tx=function(){var a=this.o.C.id,b=g.Za(this.u,function(c){return"REMOTE_CONTROL"==c.type&&c.id!=a});
-return b?b.id:""};
-g.h.Yk=function(){var a=this.o;return a.u.isActive()?a.u.u-(0,g.G)():NaN};
-g.h.UE=function(){return this.A||"UNSUPPORTED"};
-g.h.VE=function(){return this.C||""};
-g.h.sP=function(){if(!isNaN(this.Yk())){var a=this.o.u;g.Bn(a.o);a.start()}};
-g.h.vO=function(a,b){l8(this,a,b);zxa(this)};
-f8.prototype.subscribe=f8.prototype.subscribe;f8.prototype.unsubscribeByKey=f8.prototype.uh;f8.prototype.getProxyState=f8.prototype.gF;f8.prototype.disconnect=f8.prototype.ij;f8.prototype.getPlayerContextData=f8.prototype.fF;f8.prototype.setPlayerContextData=f8.prototype.AO;f8.prototype.getOtherConnectedRemoteId=f8.prototype.tx;f8.prototype.getReconnectTimeout=f8.prototype.Yk;f8.prototype.getAutoplayMode=f8.prototype.UE;f8.prototype.getAutoplayVideoId=f8.prototype.VE;f8.prototype.reconnect=f8.prototype.sP;
-f8.prototype.sendMessage=f8.prototype.vO;g.r(m8,m7);g.h=m8.prototype;g.h.le=function(a){return this.Bd.$_gs(a)};
-g.h.contains=function(a){return!!this.Bd.$_c(a)};
-g.h.get=function(a){return this.Bd.$_g(a)};
-g.h.start=function(){this.Bd.$_st()};
-g.h.add=function(a,b,c){this.Bd.$_a(a,b,c)};
-g.h.remove=function(a,b,c){this.Bd.$_r(a,b,c)};
-g.h.Zp=function(a,b,c,d){this.Bd.$_un(a,b,c,d)};
-g.h.X=function(){for(var a=0,b=this.u.length;a<b;++a)this.Bd.$_ubk(this.u[a]);this.u.length=0;this.Bd=null;m7.prototype.X.call(this)};
-g.h.wP=function(){this.P("screenChange")};
-g.h.WL=function(){this.P("onlineScreenChange")};
-v7.prototype.$_st=v7.prototype.start;v7.prototype.$_gspc=v7.prototype.xP;v7.prototype.$_gsppc=v7.prototype.cD;v7.prototype.$_c=v7.prototype.contains;v7.prototype.$_g=v7.prototype.get;v7.prototype.$_a=v7.prototype.add;v7.prototype.$_un=v7.prototype.Zp;v7.prototype.$_r=v7.prototype.remove;v7.prototype.$_gs=v7.prototype.le;v7.prototype.$_gos=v7.prototype.bD;v7.prototype.$_s=v7.prototype.subscribe;v7.prototype.$_ubk=v7.prototype.uh;var F8=null,E8=!1,n8=null,o8=null,D8=null,s8=[];g.r(G8,g.A);g.h=G8.prototype;
-g.h.X=function(){g.A.prototype.X.call(this);this.A.stop();this.C.stop();this.M.stop();this.T();this.o.unsubscribe("proxyStateChange",this.YA,this);this.o.unsubscribe("remotePlayerChange",this.km,this);this.o.unsubscribe("remoteQueueChange",this.pp,this);this.o.unsubscribe("autoplayUpNext",this.oA,this);this.o.unsubscribe("previousNextChange",this.UA,this);this.o.unsubscribe("nowAutoplaying",this.PA,this);this.o.unsubscribe("autoplayDismissed",this.nA,this);this.o=this.B=null};
-g.h.mz=function(a,b){for(var c=[],d=1;d<arguments.length;++d)c[d-1]=arguments[d];if(2!=this.o.o)if(J8(this)){if(!$7(this.o).isAdPlaying()||"control_seek"!=a)switch(a){case "control_toggle_play_pause":$7(this.o).Ya()?this.o.pause():this.o.play();break;case "control_play":this.o.play();break;case "control_pause":this.o.pause();break;case "control_seek":this.J.Cx(c[0],c[1]);break;case "control_subtitles_set_track":K8(this,c[0]);break;case "control_set_audio_track":c=c[0],J8(this)&&this.o.setAudioTrack(this.u.getVideoData(1).videoId,
-c)}}else switch(a){case "control_toggle_play_pause":case "control_play":case "control_pause":c=this.u.getCurrentTime();L8(this,0==c?void 0:c);break;case "control_seek":L8(this,c[0]);break;case "control_subtitles_set_track":K8(this,c[0]);break;case "control_set_audio_track":c=c[0],J8(this)&&this.o.setAudioTrack(this.u.getVideoData(1).videoId,c)}};
-g.h.yK=function(a){this.M.nF(a)};
-g.h.XO=function(a){this.mz("control_subtitles_set_track",g.Mb(a)?null:a)};
-g.h.pC=function(){var a=this.u.getOption("captions","track");g.Mb(a)||K8(this,a)};
-g.h.DA=function(a){if(J8(this)){this.o.unsubscribe("remotePlayerChange",this.km,this);var b=Math.round(a.volume);a=!!a.muted;var c=$7(this.o);if(b!=c.volume||a!=c.muted)this.o.setVolume(b,a),this.V.start();this.o.subscribe("remotePlayerChange",this.km,this)}};
-g.h.uL=function(){g.Mb(this.F)||Uxa(this,this.F);this.G=!1};
-g.h.YA=function(a,b){this.C.stop();2==b&&this.oC()};
-g.h.km=function(){if(J8(this)){this.A.stop();var a=$7(this.o);switch(a.playerState){case 1080:case 1081:case 1084:case 1085:this.B.H=1;break;case 1082:case 1083:this.B.H=0;break;default:this.B.H=-1}switch(a.playerState){case 1081:case 1:I8(this,new g.MC(8));this.nC();break;case 1085:case 3:I8(this,new g.MC(9));break;case 1083:case 0:I8(this,new g.MC(2));this.J.stop();H8(this,this.u.getVideoData().lengthSeconds);break;case 1084:I8(this,new g.MC(4));break;case 2:I8(this,new g.MC(4));H8(this,V7(a));
-break;case -1:I8(this,new g.MC(64));break;case -1E3:I8(this,new g.MC(128,{errorCode:"mdx.remoteerror",errorMessage:"This video is not available for remote playback."}))}a=$7(this.o).o;var b=this.F;(a||b?a&&b&&a.trackName==b.trackName&&a.languageCode==b.languageCode&&a.languageName==b.languageName&&a.kind==b.kind:1)||(this.F=a,Uxa(this,a));a=$7(this.o);-1==a.volume||Math.round(this.u.getVolume())==a.volume&&this.u.isMuted()==a.muted||this.V.isActive()||this.OC()}else Txa(this)};
-g.h.UA=function(){this.u.P("mdxpreviousnextchange")};
-g.h.pp=function(){J8(this)||Txa(this)};
-g.h.iE=function(){this.o.Hw()};
-g.h.oA=function(){};
-g.h.PA=function(a){isNaN(a)||this.u.P("mdxnowautoplaying",a)};
-g.h.nA=function(){this.u.P("mdxautoplaycanceled")};
-g.h.hI=function(a,b){-1==$7(this.o).playerState?L8(this,a):b&&this.o.seekTo(a)};
-g.h.OC=function(){if(J8(this)){var a=$7(this.o);this.w.Va(this.R);a.muted?this.u.mute():this.u.unMute();this.u.setVolume(a.volume);this.R=this.w.L(this.u,"onVolumeChange",this.DA)}};
-g.h.nC=function(){this.A.stop();if(!this.o.ga()){var a=$7(this.o);a.Ya()&&I8(this,new g.MC(8));H8(this,V7(a));this.A.start()}};
-g.h.oC=function(){this.C.stop();this.A.stop();var a=this.o.A.getReconnectTimeout();2==this.o.o&&!isNaN(a)&&this.C.start()};g.r(M8,g.Q);M8.prototype.Yb=function(){this.u.show()};
-M8.prototype.w=function(){c6("mdx-privacy-popup-cancel");this.u.hide()};
-M8.prototype.B=function(){c6("mdx-privacy-popup-confirm");this.u.hide()};g.r(N8,g.Q);N8.prototype.A=function(a){Wxa(this,a.state)};g.r(O8,g.oO);O8.prototype.J=function(){var a=this.u.getOption("remote","receivers");a&&1<a.length&&!this.u.getOption("remote","quickCast")?(this.C=g.vb(a,this.B,this),g.qO(this,(0,g.Ec)(a,this.B)),a=this.u.getOption("remote","currentReceiver"),this.Ee(this.B(a)),this.enable(!0)):this.enable(!1)};
-O8.prototype.B=function(a){return a.key};
-O8.prototype.Ye=function(a){return"cast-selector-receiver"==a?"Cast...":this.C[a].name};
-O8.prototype.gd=function(a){g.oO.prototype.gd.call(this,a);this.u.setOption("remote","currentReceiver",this.C[a]);this.w.Za()};g.r(P8,g.sL);g.h=P8.prototype;
-g.h.create=function(){Nxa(g.Oy(this.player.N()));this.A.push(g.So("yt-remote-before-disconnect",this.vK,this));this.A.push(g.So("yt-remote-connection-change",this.BM,this));this.A.push(g.So("yt-remote-receiver-availability-change",this.WA,this));this.A.push(g.So("yt-remote-auto-connect",this.zM,this));this.A.push(g.So("yt-remote-receiver-resumed",this.yM,this));this.A.push(g.So("mdx-privacy-popup-confirm",this.ZN,this));this.A.push(g.So("mdx-privacy-popup-cancel",this.YN,this));this.WA()};
-g.h.load=function(){this.player.cancelPlayback();g.sL.prototype.load.call(this);this.B=new G8(this,this.player,this.o);var a=(a=Sxa())?a.currentTime:0;var b=C8()?new Z7(x8(),void 0):null;0==a&&b&&(a=V7($7(b)));0!=a&&(this.V=a||0,this.player.P("progresssync",a,void 0));Vxa(this,this.F,this.F);g.XS(this.player.app,6)};
-g.h.unload=function(){this.player.P("mdxautoplaycanceled");this.w=this.u;g.Ke(this.B,this.o);this.o=this.B=null;g.sL.prototype.unload.call(this);g.XS(this.player.app,5);Q8(this)};
-g.h.X=function(){g.To(this.A);g.sL.prototype.X.call(this)};
-g.h.Hl=function(a,b){for(var c=[],d=1;d<arguments.length;++d)c[d-1]=arguments[d];this.loaded&&this.B.mz.apply(this.B,[a].concat(g.na(c)))};
-g.h.getAdState=function(){return this.H};
-g.h.iF=function(){return this.loaded?this.B.suggestion:null};
-g.h.Iy=function(){return this.o?$7(this.o).J:!1};
-g.h.hasNext=function(){return this.o?$7(this.o).hasNext:!1};
-g.h.getCurrentTime=function(){return this.V};
-g.h.getProgressState=function(){var a=$7(this.o),b=this.player.getVideoData();return{allowSeeking:g.R(this.player.N().experiments,"web_player_mdx_allow_seeking_change_killswitch")?this.player.Pc():!a.isAdPlaying()&&this.player.Pc(),clipEnd:b.clipEnd,clipStart:b.clipStart,current:this.getCurrentTime(),displayedStart:-1,duration:a.getDuration(),ingestionTime:a.w?a.A+T7(a):a.A,isAtLiveHead:1>=(a.w?a.u+T7(a):a.u)-this.getCurrentTime(),loaded:a.H,seekableEnd:a.w?a.u+T7(a):a.u,seekableStart:0<a.B?a.B+T7(a):
-a.B}};
-g.h.nextVideo=function(){this.o&&this.o.nextVideo()};
-g.h.DG=function(){this.o&&this.o.ZC()};
-g.h.vK=function(a){1==a&&(this.R=this.o?$7(this.o):null)};
-g.h.BM=function(){var a=C8()?new Z7(x8(),void 0):null;if(a){var b=this.w;this.loaded&&this.unload();this.o=a;this.R=null;b.key!=this.u.key&&(this.w=b,this.load())}else g.Je(this.o),this.o=null,this.loaded&&(this.unload(),(a=this.R)&&a.videoId==this.player.getVideoData().videoId&&this.player.cueVideoById(a.videoId,V7(a)));this.player.P("videodatachange","newdata",this.player.getVideoData(),3)};
-g.h.WA=function(){this.J=[this.u].concat(Pxa());var a=y8()||this.u;R8(this,a);this.player.ma("onMdxReceiversChange")};
-g.h.RN=function(a){!this.M&&g.ZC(a,8)&&(this.player.pauseVideo(),Q8(this))};
-g.h.zM=function(){var a=y8();R8(this,a)};
-g.h.yM=function(){this.w=y8()};
-g.h.ZN=function(){this.M=!0;Q8(this);E8=!1;F8&&A8(F8,1);F8=null};
-g.h.YN=function(){this.M=!1;Q8(this);R8(this,this.u);this.w=this.u;E8=!1;F8=null;this.player.playVideo()};
-g.h.Nc=function(a,b){switch(a){case "casting":return this.loaded;case "receivers":return this.J;case "currentReceiver":return b&&("cast-selector-receiver"==b.key?P7():R8(this,b)),this.loaded?this.w:this.u;case "quickCast":return 2==this.J.length&&"cast-selector-receiver"==this.J[1].key?(b&&P7(),!0):!1}};
-g.h.EG=function(){d8(this.o,"sendDebugCommand",{debugCommand:"stats4nerds "})};
-g.h.Mf=function(){return!1};
-g.h.getOptions=function(){return["casting","receivers","currentReceiver","quickCast"]};g.HL.remote=P8;})(_yt_player);
