+++ start-local.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 cd frontend/
 
 yarn build
